### 2024-09-25 #7281 @kthyng
@LC-SAR can you try to fix the compilation error?
