+++ paper/paper.bib
@@ -7,29 +7,12 @@
   doi = {10.1109/LGRS.2024.3376992}
 }
 
-
-@article{Chen2018,
-author = {Yu, Chen and Li, Zhenhong and Penna, Nigel T. and Crippa, Paola},
-title = {Generic Atmospheric Correction Model for Interferometric Synthetic Aperture Radar Observations},
-journal = {Journal of Geophysical Research: Solid Earth},
-volume = {123},
-number = {10},
-pages = {9202-9222},
-keywords = {InSAR, atmospheric correction, ECMWF, GPS, GACOS, tropospheric delay},
-doi = {https://doi.org/10.1029/2017JB015305},
-url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2017JB015305},
-eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2017JB015305},
-year = {2018}
-}
-
 @article{Quegan2001,
 author = {Quegan, Shaun and Yu, Jiong Jiong},
 doi = {10.1109/36.964973},
-file = {:home/anurag/Downloads/Filtering of multichannel SAR images.pdf:pdf},
 issn = {01962892},
 journal = {IEEE Transactions on Geoscience and Remote Sensing},
 keywords = {Agriculture,Foresty,Multitemporal filtering,Speckle reduction},
-mendeley-groups = {PhDProject,SAR/Filtering},
 number = {11},
 pages = {2373--2379},
 title = {{Filtering of multichannel SAR images}},
@@ -37,212 +20,6 @@ volume = {39},
 year = {2001}
 }
 
-@article{vasile2006intensity,
-  title={Intensity-driven adaptive-neighborhood technique for polarimetric and interferometric {SAR} parameters estimation},
-  author={Vasile, Gabriel and Trouv{\'e}, Emmanuel and Lee, Jong-Sen and Buzuloiu, Vasile},
-  journal={IEEE Transactions on Geoscience and Remote Sensing},
-  volume={44},
-  number={6},
-  pages={1609--1621},
-  year={2006},
-  doi = {10.1109/TGRS.2005.864142},
-  publisher={IEEE}
-}
-@article{lee2008improved,
-  title={Improved sigma filter for speckle filtering of {SAR} imagery},
-  author={Lee, Jong-Sen and Wen, Jen-Hung and Ainsworth, Thomas L and Chen, Kun-Shan and Chen, Abel J},
-  journal={IEEE Transactions on Geoscience and Remote Sensing},
-  volume={47},
-  number={1},
-  pages={202--213},
-  year={2008},
-  doi = {10.1109/TGRS.2008.2002881},
-  publisher={IEEE}
-}
-
-@article{chang2023extraction,
-  title={Extraction and Analysis of Radar Scatterer Attributes for {PAZ SAR} by Combining Time Series {InSAR, PolSAR,} and Land Use Measurements},
-  author={Chang, Ling and Kulshrestha, Anurag and Zhang, Bin and Zhang, Xu},
-  journal={Remote Sensing},
-  volume={15},
-  number={6},
-  pages={1571},
-  year={2023},
-  doi = {10.3390/rs15061571},
-  publisher={MDPI}
-}
-
-@article{morishita2020licsbas,
-  title={{LiCSBAS: An} open-source {InSAR} time series analysis package integrated with the {LiCSAR automated Sentinel-1 InSAR} processor},
-  author={Morishita, Yu and Lazecky, Milan and Wright, Tim J and Weiss, Jonathan R and Elliott, John R and Hooper, Andy},
-  journal={Remote Sensing},
-  volume={12},
-  number={3},
-  pages={424},
-  year={2020},
-  publisher={MDPI}
-}
-
-
-@article{Kaminski2010,
[…]
