### 2024-09-26 7de678d
Update paper.bib
