+++ paper/paper.bib
@@ -48,12 +48,14 @@ year = {2001}
   author={Curlander, J.C. and McDonough, R.N.},
   volume={11},
   year={1991},
+  ISBN = {9780471857709},
   publisher={Wiley, New York}
 }
 @book{lee2017polarimetric,
   title={Polarimetric radar imaging: from basics to applications},
   author={Lee, J.S. and Pottier, E.},
   year={2017},
+  doi = {10.1201/9781420054989},
   publisher={CRC press}
 }
 @article{long2021creating,
@@ -95,6 +97,7 @@ year = {2001}
   author={Ma, Yaobin and Ke, Peng and Hossein, A. and Chang, Ling and Wei, Jingbo},
   journal={IEEE Transactions on Geoscience and Remote Sensing},
   year={2024},
+  doi = {10.1109/TGRS.2024.3419083}
   publisher={IEEE}
 }
 
@@ -105,5 +108,6 @@ year = {2001}
   volume={15},
   pages={4559--4570},
   year={2022},
+  doi = {10.1109/JSTARS.2022.3180994},
   publisher={IEEE}
 }
