### 2024-09-26 aca3070
Update paper.bib
