+++ paper/paper.bib
@@ -97,7 +97,7 @@ year = {2001}
   author={Ma, Yaobin and Ke, Peng and Hossein, A. and Chang, Ling and Wei, Jingbo},
   journal={IEEE Transactions on Geoscience and Remote Sensing},
   year={2024},
-  doi = {10.1109/TGRS.2024.3419083}
+  doi = {10.1109/TGRS.2024.3419083},
   publisher={IEEE}
 }
 
