+++ .github/workflows/draft_pdf.yml
@@ -13,6 +13,7 @@ jobs:
           journal: joss
           # This should be the path to the paper within your repo.
           paper-path: paper/paper.md
+          branch: main
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
@@ -21,3 +22,4 @@ jobs:
           # PDF. Note, this should be the same directory as the input
           # paper.md
           path: paper/paper.pdf
+          branck: main
