+++ .github/workflows/draft_pdf.yml
@@ -22,4 +22,4 @@ jobs:
           # PDF. Note, this should be the same directory as the input
           # paper.md
           path: paper/paper.pdf
-          branck: main
+          branch: main
