+++ .github/workflows/draft_pdf.yml
@@ -13,7 +13,6 @@ jobs:
           journal: joss
           # This should be the path to the paper within your repo.
           paper-path: paper/paper.md
-          branch: main
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
@@ -22,4 +21,4 @@ jobs:
           # PDF. Note, this should be the same directory as the input
           # paper.md
           path: paper/paper.pdf
-          branch: main
+   
