### 2024-09-26 #7281 @LC-SAR
> **Submitting author:** @LC-SAR ([Ling Chang](http://orcid.org/0000-0001-8212-7221)) **Repository:** https://github.com/AlignSAR/alignSAR **Branch with paper.md** (empty if default branch): https://github.com/AlignSAR/alignSAR/tree/main/paper/paper.md **Version:** v1.1 **Editor:** Pending **Reviewers:** Pending **Managing EiC:** Kristen Thyng
>
> ## Status
> [![status](https://camo.githubusercontent.com/77d0fa6ba609bab8a03ebea7c844041bc4453fb4e7d590811326d79d321c3b67/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f30636132303137346630303433313964343365643837353666386534386638662f7374617475732e737667)](https://joss.theoj.org/papers/0ca20174f004319d43ed8756f8e48f8f)
>
> Status badge code:
>
> ```
> ```
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS @LC-SAR. **Currently, there isn't a JOSS editor assigned** to your paper.
>
> @LC-SAR if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
>
> **Editor instructions**
>
> The JOSS submission bot @editorialbot is here to help you find and assign reviewers and start the main review. To find out what @editorialbot can do for you type:
>
> ```
> @editorialbot commands
> ```

Dear Editor, Thanks for reviewing our paper. I have updated my paper.bib (https://github.com/AlignSAR/alignSAR/blob/main/paper/paper.bib) and added the missing doi for all the articles (expecting a book published in 1991 which does not have doi).  I have no problem generating the pdf by running https://github.com/AlignSAR/alignSAR/blob/main/.github/workflows/draft_pdf.yml. But when I ran '@editorialbot generate pdf', I encountered the same issue as you identified. I suspect there should be because when submitting, I gave the absolute path for 'Branch with paper.md', like this '**Branch with paper.md** (empty if default branch): https://github.com/AlignSAR/alignSAR/tree/main/paper/paper.md'.
Could you please help me remove 'https://github.com/AlignSAR/alignSAR/tree/main/paper/paper.md' for **Branch with paper.md** (empty if default branch). I think if we leave **Branch with paper.md** as empty, bot can generate the pdf. Thank you.
