### 2024-09-27 #7281 @LC-SAR
> @LC-SAR Sorry I missed that!
>
> Hi and thanks for your submission! I am looking for some specific items to make sure your submission fits our requirements at a high level (not at the more detailed review level) before moving on to finding an editor or putting this on our waitlist if no relevant editors are available. I'll comment over time as I have a chance to go through them:
>
> * [ ]  [OSI-approved](https://opensource.org/licenses) license
> * [ ]   Installation instructions
> * [ ]   Documentation exists at all
> * [ ]  Tests of some sort
> * [ ]  Check paper (make sure compiles, read through, check for length [250-1000 words](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain))
> * [ ]   [Substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
> * [ ]  Clear research application
>
> In the meantime, please take a look at the comments above ⬆️ from the editorialbot to address any DOI, license, or paper issues if you're able (there may not be any), or suggest reviewers. For reviewers, please suggest 5 reviewers from the [database listed above](https://reviewers.joss.theoj.org/lookup) or your own ([non-conflicted](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)) extended network. Their github handles are most useful to receive but please don't use "@" to reference them since it will prematurely ping them.

Thank you Kristen for your help. Our AlignSAR toolbox follows Apache-2.0 license; Installation instruction is introduced in AlignSAR_tutorial.pdf (https://github.com/AlignSAR/alignSAR/blob/main/AlignSAR_tutorial.pdf), see Section 1.1.; Documentation can be found in https://github.com/AlignSAR/alignSAR/tree/main; demonstration can also found in AlignSAR_tutorial.pdf; Submitted paper has 965 words. Considering research fields in earth science, radar remote sensing, and machine learning, five potential reviewers could be: GitHub names: [weiji14](https://github.com/weiji14), [jiangyuinsar](https://github.com/jiangyuinsar), [rogerkuou](https://github.com/rogerkuou), [yotarazona](https://github.com/yotarazona), [talha-q](https://github.com/talha-q)
