### 2024-09-28 1464900
Update input_S1_files.txt
updated for time period 2020/06 - 2024/06, only for bursts within requested polygon over Campi Flegrei, that is bursts ID '22_IW1_22892', '22_IW1_22919', '22_IW1_22947'
