+++ TS_dataset/input_S1_files.txt
@@ -1,728 +1,194 @@
-S1A_IW_SLC__1SDV_20170605T051122_20170605T051149_016894_01C1AC_08EC
-S1A_IW_SLC__1SDV_20170605T051147_20170605T051215_016894_01C1AC_00F1
-S1A_IW_SLC__1SDV_20170617T051123_20170617T051150_017069_01C711_F8B7
-S1A_IW_SLC__1SDV_20170617T051148_20170617T051216_017069_01C711_06CD
-S1A_IW_SLC__1SDV_20170629T051124_20170629T051151_017244_01CC55_9F62
-S1A_IW_SLC__1SDV_20170629T051149_20170629T051216_017244_01CC55_F948
-S1A_IW_SLC__1SDV_20170711T051124_20170711T051151_017419_01D1A6_E59E
-S1A_IW_SLC__1SDV_20170711T051149_20170711T051217_017419_01D1A6_4AAF
-S1A_IW_SLC__1SDV_20170723T051125_20170723T051152_017594_01D6EE_43A6
-S1A_IW_SLC__1SDV_20170723T051150_20170723T051218_017594_01D6EE_1AE6
-S1A_IW_SLC__1SDV_20170804T051126_20170804T051153_017769_01DC50_FF59
-S1A_IW_SLC__1SDV_20170804T051151_20170804T051219_017769_01DC50_A93C
-S1A_IW_SLC__1SDV_20170816T051126_20170816T051153_017944_01E19D_D3F1
-S1A_IW_SLC__1SDV_20170816T051151_20170816T051219_017944_01E19D_09A4
-S1A_IW_SLC__1SDV_20170828T051127_20170828T051154_018119_01E6E9_4614
-S1A_IW_SLC__1SDV_20170828T051152_20170828T051220_018119_01E6E9_9156
-S1A_IW_SLC__1SDV_20170909T051127_20170909T051154_018294_01EC47_D4B7
-S1A_IW_SLC__1SDV_20170909T051152_20170909T051220_018294_01EC47_BC9B
-S1A_IW_SLC__1SDV_20170921T051128_20170921T051155_018469_01F1A9_8443
-S1A_IW_SLC__1SDV_20170921T051153_20170921T051221_018469_01F1A9_6C50
-S1A_IW_SLC__1SDV_20171003T051153_20171003T051221_018644_01F704_8725
-S1A_IW_SLC__1SDV_20171015T051128_20171015T051155_018819_01FC5A_A681
-S1A_IW_SLC__1SDV_20171015T051153_20171015T051221_018819_01FC5A_ADF9
-S1A_IW_SLC__1SDV_20171027T051128_20171027T051155_018994_0201AC_F176
-S1A_IW_SLC__1SDV_20171027T051153_20171027T051221_018994_0201AC_D05C
-S1A_IW_SLC__1SDV_20171108T051128_20171108T051155_019169_020711_5834
-S1A_IW_SLC__1SDV_20171108T051153_20171108T051221_019169_020711_9A11
-S1A_IW_SLC__1SDV_20171120T051128_20171120T051155_019344_020C93_0CB4
-S1A_IW_SLC__1SDV_20171120T051153_20171120T051221_019344_020C93_DB0E
-S1A_IW_SLC__1SDV_20171202T051127_20171202T051154_019519_021209_87BF
-S1A_IW_SLC__1SDV_20171202T051152_20171202T051220_019519_021209_447D
-S1A_IW_SLC__1SDV_20171214T051127_20171214T051154_019694_021783_DE32
-S1A_IW_SLC__1SDV_20171214T051152_20171214T051220_019694_021783_9B2C
[…]
