### 2024-09-30 7e41919
Update batch_LiCSBAS.sh
