+++ DockerFile
@@ -19,6 +19,8 @@ RUN  cd /opt && git clone https://github.com/comet-licsar/LiCSBAS && \
 RUN git clone https://github.com/comet-licsar/licsar_proc && \
   mamba install -c conda-forge -y --file licsar_proc/requirements.txt && \
   mamba install -c conda-forge -y gmt && \
+# install LiCSBAS addons as part of licsar_extra:
+pip install git+https://github.com/comet-licsar/licsar_extra.git
   LiCSARpath="/opt/licsar_proc" && \
   echo "export LiCSARpath=/opt/licsar_proc" >> ~/.bashrc && \
   echo "export PATH=\$PATH:\$LiCSARpath/bin:\$LiCSARpath/bin/orig:\$LiCSARpath/bin/scripts:\$LiCSARpath/python" >> ~/.bashrc && \
