### 2024-10-02 #7281 @kthyng
@LC-SAR some high level comments:
* something like readthedocs (hosted docs) or using your already-existing website for more docs would be more user-friendly rather than the tutorials being a bit buried on github (for both the pdf tutorial and the notebook)
* I see your installation is docker-based. Can a user install without using docker? It is difficult for a user to contribute to development of an open source tool if installation is exclusively docker-based, with some exceptions.
* What tests do you have so users know their code is working correctly?
* Your inline and parenthetical references aren't setup correctly. Take a look at the syntax here: https://joss.readthedocs.io/en/latest/example_paper.html
