+++ paper/paper.md
@@ -37,17 +37,17 @@ date: 27 Aug 2024
 bibliography: paper.bib
 ---
 
-## Summary
+# Summary
 
-Benchmark datasets are essential for the development, evaluation, and comparison of machine learning models. However, in the field of Synthetic Aperture Radar (SAR) remote sensing [@curlander1991synthetic], there is a notable lack of large and standardized benchmark datasets, as reported by [@long2021creating], [@zhu2021deep]. Moreover, there is currently no open reference library of methods and tools for creating SAR benchmark datasets. To address these gaps, we developed AlignSAR, a comprehensive toolbox designed to create benchmark datasets specifically from SAR imagery, optimized for machine learning applications.
+Benchmark datasets are essential for the development, evaluation, and comparison of machine learning models. However, in the field of Synthetic Aperture Radar (SAR) remote sensing [@curlander1991synthetic], there is a notable lack of large and standardized benchmark datasets, as reported by [@long2021creating]; [@zhu2021deep]. Moreover, there is currently no open reference library of methods and tools for creating SAR benchmark datasets. To address these gaps, we developed AlignSAR, a comprehensive toolbox designed to create benchmark datasets specifically from SAR imagery, optimized for machine learning applications.
 
 AlignSAR enables users to process both legacy and contemporary SAR datasets, including those from Envisat, Radarsat, and Sentinel-1, facilitating the extraction of representative SAR signatures. Additionally, the toolbox supports the alignment of other geospatial observations, such as LiDAR and optical satellite data, onto SAR reference grids, thereby enriching each SAR pixel with additional signatures. The SAR benchmark datasets created with AlignSAR, constructed from these enriched and aligned signatures, are then ready for use in a variety of machine-learning tasks and can be shared easily on the Cloud, e.g. by using STAC (SpatioTemporal Asset Catalogs).
 
-## Statement of Need
+# Statement of Need
 
-SAR techniques, including InSAR (Interferometric SAR) [@bamler1998synthetic] and PolSAR (Polarimetric SAR) [@lee2017polarimetric], are well-established and are employed in natural and anthropogenic hazard monitoring, as well as land use and land cover classification. These techniques typically employ conventional methods such as persistent scatterer interferometry and the small baseline subset approach [@ferretti2001permanent], [@lanari2004small], recently in conjunction with machine learning techniques. A variety of commercial and open-source processing tools have been developed to support different SAR applications. For instance, The ESA SentiNel Application Platform  [SNAP](https://earth.esa.int/eogateway/tools/snap), a JAVA-based tool, which contains the Sentinel-1 Toolbox that is commonly used for standard InSAR and PolSAR decomposition analysis. [Doris](https://github.com/TUDelftGeodesy/Doris), stands for Delft object-oriented radar interferometric software, and is another open-source tool designed for InSAR analysis, though it is somewhat outdated and has been partially deprecated. On the commercial side, [Gamma](https://www.gamma-rs.ch/software) offers a robust solution for InSAR analysis, also integrated into platforms like [ASF Services – On Demand Processing](https://asf.alaska.edu/asf-services-hyp3-processing/) and [COMET LiCSAR System](https://comet.nerc.ac.uk/comet-lics-portal/). 
[…]
