### 2024-10-03 #7281 @LC-SAR
> @LC-SAR some high level comments:
>
> * something like readthedocs (hosted docs) or using your already-existing website for more docs would be more user-friendly rather than the tutorials being a bit buried on github (for both the pdf tutorial and the notebook)
> * I see your installation is docker-based. Can a user install without using docker? It is difficult for a user to contribute to development of an open source tool if installation is exclusively docker-based, with some exceptions.
> * What tests do you have so users know their code is working correctly?
> * Your inline and parenthetical references aren't setup correctly. Take a look at the syntax here: https://joss.readthedocs.io/en/latest/example_paper.html

Dear Kristen,
please see my response to your four points below:

- For every functionality, we also have a brief description in its gihub folder. For intance, for the modified Doris software, in the folder 'itd-doris_5_patch2023', https://github.com/AlignSAR/alignSAR/tree/main/itc-doris_5_patch2023, we provide its description on the github website.
- Docker is one option to run our tool. We also describe how to install all required software manually on the user's local computer and list troubleshooting. For instance, the guideline to install Doris, see https://github.com/AlignSAR/alignSAR/blob/main/AlignSAR_tutorial.pdf, Appendix A (page 30).
- We have offered a set of sample data (NetCDF), see https://github.com/AlignSAR/alignSAR/tree/main/jupyter_notebook_demo, which allows the users to test without creating SAR signatures from raw SAR (Single-Look-complex) data. More sample data have been uploaded on EOTDL: https://www.eotdl.com/datasets/Alignsar/, https://www.eotdl.com/datasets/AlignSAR-Groningen-Sentinel1-Q0/ by us.
- I have modified the paper.md (https://github.com/AlignSAR/alignSAR/blob/main/paper/paper.md), for instance, changed '## Summary' to '# Summary', updated Figure index and caption, changed reference from [@zhu2021deep], [ @kulshrestha2022use] to [@zhu2021deep; @kulshrestha2022use]. I tried to add the year like [@zhu2021deep:2021], but compiling pdf failed. So I excluded the year while the reference format in the new pdf is the same as JOSS paper format. Hope it suffices.
