### 2024-10-12 c8846ad
Update batch_LiCSBAS.sh
setting correct download start-end dates
