### 2024-10-30 #7281 @LC-SAR
Dear @kthyng Kristen,
It is about a month since our last contact. I wonder whether you have any further questions regarding my response to your points (see above).
Thanks and with regards,
Ling
