### 2024-10-30 #7281 @kthyng
Please check that your inline vs. parenthetical references are correct. You can compare with the example paper here: https://joss.readthedocs.io/en/latest/example_paper.html. For example, in line 11 you have references in parentheses that shouldn't be.
