+++ paper/paper.md
@@ -65,7 +65,7 @@ AlignSAR is available as an Apache-2.0-licensed open-source toolbox hosted on [G
 
 2. [jupyter_notebook_demo](https://github.com/AlignSAR/alignSAR/tree/main/jupyter_notebook_demo): a jupyter notebook to demonstrate how to extract, visualize and analyse SAR signatures, with the sample output generated by Doris.
 
-3. [rdcode](https://github.com/AlignSAR/alignSAR/tree/main/rdrcode): the radarcoding scripts to align other geospatial data onto SAR reference grid. For method details refer to [@Kulshresta2023].
+3. [rdcode](https://github.com/AlignSAR/alignSAR/tree/main/rdrcode): the radarcoding scripts to align other geospatial data onto SAR reference grid. For method details refer to @Kulshresta2023.
 
 4. [snap_graphs](https://github.com/AlignSAR/alignSAR/tree/main/snap_graphs): graphs for SNAP used within the toolbox (e.g. rdr<->geocode).
 
@@ -77,7 +77,7 @@ AlignSAR is available as an Apache-2.0-licensed open-source toolbox hosted on [G
 
 8. [signature_extraction.py](https://github.com/AlignSAR/alignSAR/blob/main/signature_extraction.py): a script to extract representative SAR signatures.
 
-9. [speckle_filt.py](https://github.com/AlignSAR/alignSAR/blob/main/speckle_filt.py): a script for spatio-temoporal speckle filtering. For more details, refer to [@Quegan2001].
+9. [speckle_filt.py](https://github.com/AlignSAR/alignSAR/blob/main/speckle_filt.py): a script for spatio-temoporal speckle filtering. For more details, refer to @Quegan2001.
 
 10. [MLscripts](https://github.com/AlignSAR/alignSAR/tree/main/MLscripts): scripts for machine learning analysis. [Yolov8](https://github.com/ultralytics/ultralytics), ANN (Artificial Neural Network) and Siamese are separately used for Object Detection (India), Land Use Land Cover classification (Netherlands), and Change Detection (Poland). 
 
