### 2024-10-31 #7281 @LC-SAR
> **Five most similar historical JOSS papers:**
>
> [Geodata-Harvester: A Python package to jumpstart geospatial data extraction and analysis](https://github.com/openjournals/joss-reviews/issues/5205) Submitting author: `@sebhaan` Handling editor: `@hugoledoux (Active)` Reviewers: `@lukasbeuster, @martibosch` Similarity score: 0.6934
>
> [Synaptus: A Matlab/Octave toolbox for synthetic aperture ultrasound imaging](https://github.com/openjournals/joss-reviews/issues/4185) Submitting author: `@mh-skjelvareid` Handling editor: `@Kevin-Mattheus-Moerman (Active)` Reviewers: `@emanuelhuber, @Kevin-Mattheus-Moerman` Similarity score: 0.6775
>
> [PolSAR tools: A QGIS plugin for generating SAR descriptors](https://github.com/openjournals/joss-reviews/issues/2970) Submitting author: `@Narayana-Rao` Handling editor: `@hugoledoux (Active)` Reviewers: `@liberostelios, @HenrikJanPersson` Similarity score: 0.6758
>
> [Trash AI: A Web GUI for Serverless Computer Vision Analysis of Images of Trash](https://github.com/openjournals/joss-reviews/issues/5136) Submitting author: `@wincowgerDEV` Handling editor: `@arfon (Active)` Reviewers: `@domna, @luxaritas` Similarity score: 0.6716
>
> [deeplenstronomy: A dataset simulation package for strong gravitational lensing](https://github.com/openjournals/joss-reviews/issues/2854) Submitting author: `@rmorgan10` Handling editor: `@pdebuyl (Active)` Reviewers: `@shreyasbapat, @jiwoncpark` Similarity score: 0.6660
>
> ⚠️ **Note to editors:** If these papers look like they might be a good match, click through to the review issue for that paper and invite one or more of the authors _before_ considering asking the reviewers of these papers to review again for JOSS.

Dear Kristen,
Thank you. I am afraid that I don't know which action I should take in response to this message. Is this message about proposing potential reviewers for our manuscript? If so, on 27 September, I checked the 'database' you shared and suggested five potential reviewers could be: GitHub names: [weiji14](https://github.com/weiji14), [jiangyuinsar](https://github.com/jiangyuinsar), [rogerkuou](https://github.com/rogerkuou), [yotarazona](https://github.com/yotarazona), [talha-q](https://github.com/talha-q), considering research fields in earth science, radar remote sensing, and machine learning.
