### 2024-10-31 #7281 @LC-SAR
> Sorry @LC-SAR, I will follow up now.
>
> > For every functionality, we also have a brief description in its gihub folder. For intance, for the modified Doris software, in the folder 'itd-doris_5_patch2023', https://github.com/AlignSAR/alignSAR/tree/main/itc-doris_5_patch2023, we provide its description on the github website.
>
> I don't really understand this response. Can you explain again? I am particularly curious about docs and tutorials that a user could use to start using your software and understand its functionality. The tutorial pdf is excellent, it is just a bit difficult for a user to jump into compared to more modern html like readthedocs. Also having everything in one place is less confusing.
>
> > Docker is one option to run our tool. We also describe how to install all required software manually on the user's local computer and list troubleshooting. For instance, the guideline to install Doris, see https://github.com/AlignSAR/alignSAR/blob/main/AlignSAR_tutorial.pdf, Appendix A (page 30).
>
> Is Doris all of the library, or just part of it?
>
> > We have offered a set of sample data (NetCDF), see https://github.com/AlignSAR/alignSAR/tree/main/jupyter_notebook_demo, which allows the users to test without creating SAR signatures from raw SAR (Single-Look-complex) data. More sample data have been uploaded on EOTDL: https://www.eotdl.com/datasets/Alignsar/, https://www.eotdl.com/datasets/AlignSAR-Groningen-Sentinel1-Q0/ by us.
>
> Ok, this is good. Normally there would also be unit tests to make sure the software is working correctly during development and for any users who want to contribute to development. This would probably come up during review.

Dear Kristen,

Regarding your comment '> I don't really understand this response. Can you explain again? I am particularly curious about docs and tutorials that a user could use to start using your software and understand its functionality. The tutorial pdf is excellent, it is just a bit difficult for a user to jump into compared to more modern html like readthedocs. Also having everything in one place is less confusing.', here is my further explanation:
The idea of offering this tutorial is to give a detailed and stepwise description of how to install and run our AlignSAR toolbox, the basic methodologies behind it, and the expected outputs based on a user case. I understand pdf format is not recommended in this case, so I converted this tutorial into html and shared it on Github, see https://lc-sar.github.io/alignsar_tutorial_html/. Would this html file suffice? If so, I can finalize this html file.

Regarding your question '> Is Doris all of the library, or just part of it?', Doris is one of the tools within AlignSAR toolbox. If you read our manuscript, on page 2, from line 63, we have listed all other tools besides Doris.

Thank you,

Ling
