### 2024-11-04 #7281 @kthyng
@LC-SAR Ok I reread your comments to see that you said all the software can be installed manually, not just Doris. I misunderstood.

Thank you for trying with the html vs the pdf. It looks like the formatting doesn't work out very well on the html the way it's processed so best not to do it this way.

Some of the comments in this issue are from a bot, editorialbot, and they are just to help out.
