### 2024-11-04 3596e33
Update batch_LiCSBAS.sh
