### 2024-11-21 #7281 @rwegener2
👋🏻 @josephhardinee or @Scivision, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
