### 2024-11-26 #7532 @editorialbot
[REVIEW]: AlignSAR: An open-source toolbox of SAR benchmark dataset creation for machine learning applications
**Submitting author:** @LC-SAR (<a href="http://orcid.org/0000-0001-8212-7221">Ling Chang</a>)
**Repository:** https://github.com/AlignSAR/alignSAR
**Branch with paper.md** (empty if default branch): main
**Version:** v1.1
