### 2024-11-26 #7532 @rwegener2
👋🏼 @LC-SAR, @philippemiron, and @shahchiragh this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```
