### 2024-12-17 3e614b9
Update batch_LiCSBAS.sh
