### 2024-12-20 AlignSAR/alignSAR#2 @philippemiron
JOSS Review
Hi, I'm reviewing the repo for the JOSS publication (https://github.com/openjournals/joss-reviews/issues/7532#issuecomment-2549086865), and I have some comments. I summarized them into some itemized issues.

- [x] the package `AlignSAR` as a lot of scripts and examples, but the organization of the repo is hard to follow. Would it be possible to follow a more standard structure such has:
```
alignSAR/
|─── src/
│   └── alignsar/
│        └── __init__.py
│        └── core/
│        └── utils/
│        └── mlprocessing/
|─── paper/
│   └── flowchart.png (that is in the paper)
│   └── user_guide/
|─── examples/
│    └── data/
│    └── tutorials/
│    └── notebooks/
|─── dependencies/
│    └── doris_patched2023/
│    └── stac/
|─── tests/
|─── .env
|─── .gitignore
|─── Dockerfile
|─── pyproject.toml (or a requirements.txt)
|─── README.md
|─── LICENSE
|─── zenodo.json
```
- [x] include some tests to validate the scripts and utilities
- [x] is it possible to add citation of Doris and Snap in the paper
- [x] add `__pycache__` to `.gitignore` and remove from the repo
- [x] usually the `Dockerfile` is with this orthograph. This breaks the tutorial because when you execute `docker build -t alignsar .` it is looking for a `Dockerfile`.
- [x] environment variable in `bashrc_alignsar` can be set in a `.env` file and imported in the Dockerfile
- [x] Unable to build the image
```
git clone git@github.com:AlignSAR/alignSAR.git
cd alignSAR

~/repos/alignSAR on  main
✗ docker build -f DockerFile -t alignsar .
[+] Building 0.0s (1/1) FINISHED                                                      docker:desktop-linux
 => [internal] load build definition from DockerFile                                                  0.0s
 => => transferring dockerfile: 4.38kB                                                                0.0s
DockerFile:24
--------------------
  23 |     pip install git+https://github.com/comet-licsar/licsar_extra.git
  24 | >>>   LiCSARpath="/opt/licsar_proc" && \
  25 | >>>   echo "export LiCSARpath=/opt/licsar_proc" >> ~/.bashrc && \
  26 | >>>   echo "export PATH=\$PATH:\$LiCSARpath/bin:\$LiCSARpath/bin/orig:\$LiCSARpath/bin/scripts:\$LiCSARpath/python" >> ~/.bashrc && \
  27 | >>>   echo "export PYTHONPATH=\$PYTHONPATH:\$LiCSARpath/python:\$LiCSARpath/python/LiCSAR_lib:\$LiCSARpath/python/LiCSAR_db" >> ~/.bashrc && \
  28 | >>>   echo "source activate py39" >> ~/.bashrc
  29 |
--------------------
ERROR: failed to solve: dockerfile parse error on line 24: unknown instruction: LiCSARpath="/opt/licsar_proc"
```

Once this is fixed, I can test some of the utilities and finish the review.

Cheers,
