### 2024-12-25 AlignSAR/alignSAR#3 @shahchiragh
Related to review - CShah
Hello! As part of the review process for JOSS (https://github.com/openjournals/joss-reviews/issues/7532), I’ve below comments and suggestions:

- **Codebase Organization:**
The current organization of the codebase seems unstructured, making it harder to navigate. I recommend adhering to standard best practices for organizing files. For instance, tutorials, examples, demo, paper, etc. could be moved to a dedicated module such as `examples/` or `documentation/` for better clarity and usability. Same goes for dependency, script files, etc.

- **Unwanted Files/Directories:**
Please clean or remove unnecessary files and directories to streamline the repository and avoid confusion.

- **Dockerfile Naming:**
The file [DockerFile](https://github.com/AlignSAR/alignSAR/blob/main/DockerFile) should be renamed to Dockerfile (case-sensitive) to follow Docker's naming convention. This ensures compatibility with the standard docker build command. While I was able to continue with `docker build -f DockerFile -t alignsar .`, the docker build process encountered below error:
```
$ docker build -f DockerFile -t alignsar .
[+] Building 0.0s (1/1) FINISHED                                                                                      docker:desktop-linux
 => [internal] load build definition from DockerFile                                                                                  0.0s
 => => transferring dockerfile: 4.34kB                                                                                                0.0s
DockerFile:24
--------------------
  23 |     pip install git+https://github.com/comet-licsar/licsar_extra.git
  24 | >>>   LiCSARpath="/opt/licsar_proc" && \
  25 | >>>   echo "export LiCSARpath=/opt/licsar_proc" >> ~/.bashrc && \
  26 | >>>   echo "export PATH=\$PATH:\$LiCSARpath/bin:\$LiCSARpath/bin/orig:\$LiCSARpath/bin/scripts:\$LiCSARpath/python" >> ~/.bashrc && \
  27 | >>>   echo "export PYTHONPATH=\$PYTHONPATH:\$LiCSARpath/python:\$LiCSARpath/python/LiCSAR_lib:\$LiCSARpath/python/LiCSAR_db" >> ~/.bashrc && \
  28 | >>>   echo "source activate py39" >> ~/.bashrc
  29 |
--------------------
ERROR: failed to solve: dockerfile parse error on line 24: unknown instruction: LiCSARpath="/opt/licsar_proc"
```
- **Alternate Code Installation Process:**
Since the project is primarily Python-based, it would be beneficial to offer an alternate installation process besides Docker. For example, providing a method to run the code using a virtual environment (pyenv) with a requirements.txt file would be helpful. This is not a mandatory request but could enhance accessibility for users who prefer not to use Docker.

- **Tests for Validation:**
I did not find any tests in the codebase. Adding test cases is crucial to validate the functionality of the code and ensure its reliability. I strongly recommend incorporating a testing framework and providing coverage for key components of the codebase.

These improvements will significantly enhance the quality, usability, and maintainability of the project. Let me know if you need further clarification or assistance!

Thanks!
