### 2024-12-25 #7532 @shahchiragh
Hi @rwegener2 , I’ve completed an initial round of the review. However, a few checks, specifically related to _installation_ and _documentation_, are still pending as they depend on the software installation process. I will revisit these once the authors address the comments I’ve summarized in the repo issue.

Additionally, there are a few aspects related to the software paper, such as the "Quality of Writing" and "References," that I would like to bring to the authors' attention for further refinement of the paper. Please feel free to reach out for any questions or comments. Thanks!
