+++ paper/paper.bib
@@ -111,3 +111,14 @@ year = {2001}
   doi = {10.1109/JSTARS.2022.3180994},
   publisher={IEEE}
 }
+@article{moreira2013tutorial,
+  title={A tutorial on synthetic aperture radar},
+  author={Moreira, Alberto and Prats-Iraola, Pau and Younis, Marwan and Krieger, Gerhard and Hajnsek, Irena and Papathanassiou, Konstantinos P},
+  journal={IEEE Geoscience and remote sensing magazine},
+  volume={1},
+  number={1},
+  pages={6--43},
+  year={2013},
+  doi = {10.1109/MGRS.2013.2248301},
+  publisher={IEEE}
+}
