### 2024-12-26 ff84ecd
Update paper.md
repace @curlander1991synthetic with moreira2013tutorial, line 42. because @curlander1991synthetic does not have doi
