### 2024-12-26 9d87b11
Update paper.bib
