+++ paper/paper.bib
@@ -44,7 +44,7 @@ year = {2001}
   publisher={IOP Publishing}
 }
 @book{curlander1991synthetic,
-  title={Synthetic aperture radar},
+  title={Synthetic aperture radar: systems and signal processing},
   author={Curlander, J.C. and McDonough, R.N.},
   volume={11},
   year={1991},
