### 2024-12-26 bcf7470
Update paper.bib
