+++ paper/paper.bib
@@ -122,3 +122,27 @@ year = {2001}
   doi = {10.1109/MGRS.2013.2248301},
   publisher={IEEE}
 }
+@article{kampes1999doris,
+  title={Doris: The delft object-oriented radar interferometric software},
+  author={Kampes, Bert and Usai, Stefania},
+  booktitle={Proceedings of the 2nd International Symposium on Operationalization of Remote Sensing, Enschede, The Netherlands},
+  volume={16},
+  pages={20},
+  year={1999}
+}
+@software{doris,
+  author = {Delft University of Technology},
+  title = {DORIS: The delft object-oriented radar interferometric software},
+  year = {2008},
+  url = {https://doris.tudelft.nl/},
+  note = {Version 5.0 Beta, accessed 2023-02-01}
+}
+@software{esasnap,
+  author = {{European Space Agency (ESA)}},
+  title = {{Sentinel Application Platform (SNAP)}},
+  year = {2014},
+  url = {https://step.esa.int/},
+  note = {Version 9.0, accessed 2023-02-01}
+}
+
+
