+++ paper/paper.md
@@ -49,7 +49,7 @@ SAR techniques, including InSAR (Interferometric SAR) [@bamler1998synthetic] and
 
 As an increasing number of dedicated SAR missions are launched, the community of SAR users is expanding, with growing interest in applying SAR data within machine learning frameworks [@zhu2021deep; @kulshrestha2022use; @ma2024despeckling]. Despite this trend, the aforementioned tools are not specifically designed for machine learning applications. Furthermore, directly applying methods developed for other types of signals and images to SAR data, without considering SAR representative signatures, can lead to biased and inaccurate machine learning outcomes. 
 
-AlignSAR toolbox developed new methods and tools upon existing tools such as Doris and SNAP, and enhances the functionalities, optimizes key processes, and extends the capabilities of the tools used, see the details in [Main Functionalities](#Main Functionalities). It standardizes the extraction procedure of representative SAR signatures, and provides radar- and geo-coded geospatial observations that are all aligned onto SAR reference coordinates and then associates additional signatures from these geospatial data with the corresponding SAR pixels. The SAR signatures are classified into three categories:  
+AlignSAR toolbox developed new methods and tools upon existing tools such as Doris [@doris] and SNAP [@esasnap], and enhances the functionalities, optimizes key processes, and extends the capabilities of the tools used, see the details in [Main Functionalities](#Main Functionalities). It standardizes the extraction procedure of representative SAR signatures, and provides radar- and geo-coded geospatial observations that are all aligned onto SAR reference coordinates and then associates additional signatures from these geospatial data with the corresponding SAR pixels. The SAR signatures are classified into three categories:  
 
 Category 1: Single polarimetric signatures, like amplitude, intensity, backscatter coefficient, interferometric and SLC (Single Look Complex) phase and coherence;  
 Category 2: Multi polarimetric signatures, such as co-, dual polarization cross product, and summation, difference and ratio of the co-, dual, quad polarization intensities, and entropy, scattering mechanisms, SAR vegetation index (RVI), and SAR soil moisture.  
