### 2024-12-26 de7c5a4
Update paper.bib
