+++ paper/paper.bib
@@ -131,7 +131,7 @@ year = {2001}
   year={1999}
 }
 @software{doris,
-  author = {Delft University of Technology},
+  author = {{Delft University of Technology (TUDelft)}},
   title = {DORIS: The delft object-oriented radar interferometric software},
   year = {2008},
   url = {https://doris.tudelft.nl/},
