### 2024-12-26 bc8bc13
Update paper.bib
