+++ paper/paper.bib
@@ -144,5 +144,27 @@ year = {2001}
   url = {https://step.esa.int/},
   note = {Version 9.0, accessed 2023-02-01}
 }
-
+@software{gamma,
+  author = {{GAMMA Remote Sensing}},
+  title = {{GAMMA software}},
+  year = {1995},
+  url = {https://www.gamma-rs.ch/software}
+}
+@software{asf,
+  author = {{NASA Distributed Active Archive Center (DAAC)}},
+  title = {{Alaska Satellite Facility (ASF)}},
+  year = {2017},
+  url = {https://search.asf.alaska.edu}
+}
+@article{lazecky2020licsar,
+  title={{LiCSAR: An automatic InSAR} tool for measuring and monitoring tectonic and volcanic activity},
+  author={Lazeck{\`y}, Milan and Spaans, Karsten and Gonz{\'a}lez, Pablo J and Maghsoudi, Yasser and Morishita, Yu and Albino, Fabien and Elliott, John and Greenall, Nicholas and Hatton, Emma and Hooper, Andrew and others},
+  journal={Remote Sensing},
+  volume={12},
+  number={15},
+  pages={2430},
+  year={2020},
+  doi = {10.3390/rs12152430},
+  publisher={MDPI}
+}
 
