+++ tests/MLscripts/Netherlands - LULC/data_preparation/data_preparation.py
@@ -1,0 +1,271 @@
+import rioxarray
+import matplotlib.pyplot as plt
+from netCDF4 import Dataset
+import xarray as xr
+import numpy as np
+import matplotlib.cm as cm
+import h5py
+import os
+import pandas as pd
+import collections, numpy
+import time
+
+class Labeling:
+    """This class aims to perform the data labeling, based on the probability of belonging to each class """
+    def get_data(x):
+        #This functions reads the data of the NetCDF and convert is as pandas dataframe
+        
+        # read netCDF
+        # groningen_x1 = xr.open_dataarray(x)
+        groningen_x1 = nc.Dataset(x)
+        
+        #Convert to numpy array
+        # xarr = groningen_x1.to_numpy()
+        variable_names = groningen_x1.variables.keys()
+        xarr = [np.array(groningen_x1.variables[variable_name][:]) for variable_name in variable_names]
+        xarr = np.stack(xarr, axis = -1)
+
+        #Reshape the data
+        xarr_df = np.reshape(xarr, (xarr.shape[0]*xarr.shape[1],xarr.shape[2]))[:,:14]
+        # xarr_df = np.reshape(xarr, (len(xarr[0])*len(xarr[0][0]),len(xarr)))
+        
+        #Convert numpy array to pandas dataframe
+        xdf = pd.DataFrame(xarr_df, columns=np.stack(list(groningen_x1.variables.keys()), axis = -1)[:14]
+
+        return groningen_x1,xdf
+        
+    def obtain_labels(xdf, th, a,b):
+        #This function assigns a label to each pixel
+        
+        #Cut the dataset according to a and b
+        xdf = xdf[a:b]
+        
+        #Reset index
+        xdf = xdf.reset_index()
+        
+        #Convert dataframe to numpy array
+        df_np=xdf.to_numpy()
+        
+        #Initialize necessary variables
+        df_reference = []
+        l = []
+        
+        #Only necesary to print the progress of the pixel labeling, by printing the #of pixels that
+        #have been already labeled 
+        nn = int(len(xdf)*0.05)
+        
+        """CHANGE: START"""
+        #Create one dataframe for each label
+        ### You must add the additional labels of your data
+        buildings = xdf['Buildings']
+        railways = xdf['Railways']
+        water = xdf['Water']
+        roads = xdf['Roads']
+        ##### label_5 = xdf['label_5']
+        ##### label_6 = xdf['label_6']
+        ##### ...
+        ##### label_n = xdf['label_n']
+        """CHANGE: END"""
+        
+        #For loop that goes over all pixels
+        for i in range(len(xdf)):
+            #Print the progress of the label assignation every 5%
+            if i%nn==0:
+                msg = "Loading " + str(i) + " images\n"
+                print(msg)
+            
+            """CHANGE: START"""
+            #If statements to obtain the labels for each pixel
+            if buildings[i] >= th:
+                #a_label = [1,0,0,0,0,...,0]
+                l.append(0)
+                df_reference.append(df_np[i])
+            if railways[i] >= th:
+                #a_label = [0,1,0,0,0,...,0]
+                l.append(1)
+                df_reference.append(df_np[i])
+            if water[i] >= th:
+                #a_label = [0,0,1,0,0,...,0]
+                l.append(2)
+                df_reference.append(df_np[i])
+            if roads[i] >= th:
+                #a_label = [0,0,0,1,0,...,0]
+                l.append(3)
+                df_reference.append(df_np[i])
+            ### Add here the corresponding if statements, one for each additional label you have
+            
+            ##### if label_5[i] >= th:
+            #####    #a_label = [0,0,0,1,0,...,0]
+            #####    l.append(4)
+            #####    df_reference.append(df_np[i])
+            #####
[…]
