+++ TS_dataset/batch_LiCSBAS.sh
@@ -1,483 +1,0 @@
-#!/bin/bash -eu
-
-# LiCSBAS steps:
-#  01: LiCSBAS01_get_geotiff.py
-#  02to05: LiCSBAS02to05_unwrap.py   or   02: LiCSBAS02_ml_prep.py
-#                                         03: LiCSBAS03op_GACOS.py (optional)
-#                                         04: LiCSBAS04op_mask_unw.py (optional)
-#                                         05: LiCSBAS05op_clip_unw.py (optional)
-#  11: LiCSBAS11_check_unw.py
-#  (optional) 120: LiCSBAS120_choose_reference.py   - RECOMMENDED, especially if nullification is used, thus added to cometdev
-#  12: LiCSBAS12_loop_closure.py
-#  13: LiCSBAS13_sb_inv.py
-#  14: LiCSBAS14_vel_std.py
-#  15: LiCSBAS15_mask_ts.py
-#  16: LiCSBAS16_filt_ts.py
-
-#
-# Status of COMET dev version - the experimental functions are turned on with 
-#
-#################
-### Settings ####
-#################
-start_step="01"	# 01-05, 11-16
-end_step="16"	# 01-05, 11-16
-
-cometdev='1' # will enable: p120refp, singular LS method, errors nullification
-nlook="1"	# multilook factor, used in step02
-GEOCmldir="GEOCml"${nlook}	# If start from 11 or later after doing 03-05, use e.g., GEOCml${nlook}GACOSmaskclip
-n_para="8" # Number of parallel processing in step 02-05,12,13,16. default: number of usable CPU
-gpu="n"	# y/n
-check_only="n" # y/n. If y, not run scripts and just show commands to be done
-
-logdir="log"
-log="$logdir/$(date +%Y%m%d%H%M)$(basename $0 .sh)_${start_step}_${end_step}.log"
-
-freq="" # default: 5.405e9 Hz
-
-### Running the updated pipelines:
-run_reunwrapping='y' # y/n. default: 'n'. Reunwrapping would use 02to05 script instead of the original 02[,03,04,05]
-
-### Optional steps (03-05) ###
-order_op03_05="03 04 05"	# can change order e.g., 05 03 04
-do03op_GACOS="n"	# y/n
-do04op_mask="n"	# y/n
-do05op_clip="n"	# y/n
-p04_mask_coh_thre=""	# e.g. 0.2
-p04_mask_range=""	# e.g. 10:100/20:200 (ix start from 0)
-p04_mask_range_file=""	# Name of file containing range list
-p05_clip_range=""	# e.g. 10:100/20:200 (ix start from 0)
-p05_clip_range_geo="14.02639/14.25161/40.71439/40.93961"	# e.g. 130.11/131.12/34.34/34.6 (in deg)
-
-# Optional reunwrapping:
-p02to05_freq=$freq # default: 5.405e9 Hz
-p02to05_gacos="y" # y/n. default: 'y'. Use gacos data if available (for majority of epochs, data without GACOS corr would be dropped)
-p02to05_hgtcorr="n" # y/n. default: 'n'. Recommended for regions with high and varying topography
[…]
