+++ alignsar/itc-doris_5_patch2023/doris/doris_core/Makefile
@@ -1,0 +1,255 @@
+#########################################################################
+# Makefile for the Doris software.					#
+# Created by: configure						#
+# 04-Dec-1998                                                           #
+#                                                                       #
+# Delft University of Technology                                        #
+# Delft Institute of Earth Observation and Space Systems                #
+# http://doris.tudelft.nl               				#
+# See also the user manual, annex installation.				#
+#									#
+### Usage ### 								#
+# The command: "make" creates and optimized version of doris.		#
+# The command: "make install" installs it andcleans up.			#
+# If this does not work, try a more controlled approach by:		#
+#									#
+# 0. inspect set up of this Makefile					#
+# 1. set CC DEFS LIBS etc., first set these to debug values		#
+#    (read NOTE there if problems)					#
+# 2. compile software sources to *.o:     "make swobjs"			#
+# 3. link object together to executable:  "make doris"			#
+# 4. install executable:                  "make install"		#
+# 5. remove object files:                 "make clean"			#
+#									#
+# BTW. "make" does "swobjs doris" by default.				#
+#									#
+###        								#
+# Successfully tested for compilers/platforms:				#
+#  GNU g++ version 2.7.2.2 on HP-UX B.10.20 A 9000/785			#
+#  GNU g++ version 2.95.2  on HP-UX B.10.20 A 9000/785			#
+#  HP  aCC version A.01.07 on HP-UX B.10.20 A 9000/785			#
+#  GNU g++ version ?       on Linux X86					#
+#  SGI ?								#
+#  Sun Studion 9,10,11     on Solaris Sparc and X86/AMD64		#
+#  GNU g++ versions v3.4, v4.0, v4.1, v4.2, v4.3 on Linux X86 and AMD64 # 
+#  Sun Sudion v11          on Linux AMD64				#
+#  Intel Compilers v9, v10, v11 on Linux AMD64				#
+#########################################################################
+###################################################################
+###################################################################
+# Please change if required: DEF[1-6], CC, CFLAGS, LIBDIR,      ###
+###################################################################
+### The shell used by this makefile ###
+SHELL   = /bin/sh
+
+### Specify compiler/installation directory ###
+INSTALLDIR = /usr/local/bin
+CC	   = g++
+SCRIPTSDIR = ../bin
+
+### Define statements controlling compilation ###
+# Comment out DEF[1-8] statements when appropriate.
+DEF1    = -D__DEBUGMAT1	# 		-1- range checking in matrix class
+DEF2    = -D__DEBUGMAT2	# 		-2- info matrix class (debug only)
+DEF3    = -D__DEBUG #			-3- debug for other files
+DEF4    = -Wno-deprecated            #  -4- do not display warnings due to depricated entries 
+DEF5    = -D_LARGEFILE64_SOURCE -D_FILE_OFFSET_BITS=64 -D_LARGEFILE_SOURCE #  -5- support for large files
+#DEF6    = -DNO_FASTTRIG              #  -6-  extra functional switches
+#DEF9     = -D__NO_IOS_BINARY__ # -9- comment out if you ios::binary (most will)
+#
[…]
