+++ alignsar/itc-doris_5_patch2023/doris/doris_stack/functions/precise_read.py
@@ -1,0 +1,155 @@
+# This file contains several functions to read data from precise orbit files of sentinel-1 data.
+# Scripts were created by  Wu Wenhao, Wuhan university and adapted by Gert Mulder, TU Delft
+
+import time
+import calendar
+import numpy as np
+import os, sys
+from scipy.interpolate import interp1d
+import scipy.interpolate as inter
+import calendar
+
+
+def orbit_read(input_EOF_FileName):
+
+    try:
+        import xml.etree.cElementTree as etree
+    except:
+        try:
+            from lxml import etree
+        except:
+            #import xml.etree.ElementTree as etree
+            print 'Failed to load lxml.etree or xml.etree.cElementTree'
+            sys.exit(1)
+    #print(input_EOF_FileName)
+    inTree = etree.parse(input_EOF_FileName)
+    queryList = {
+                # orbit inf
+               'Mission'  : './/Earth_Explorer_Header/Fixed_Header/Mission',\
+               'Validity_Start': './/Earth_Explorer_Header/Fixed_Header/Validity_Period/Validity_Start',\
+               'Validity_Stop': './/Earth_Explorer_Header/Fixed_Header/Validity_Period/Validity_Stop',\
+               'orbitABS' : './/Data_Block/List_of_OSVs/OSV/Absolute_Orbit',\
+               'orbitTime': './/Data_Block/List_of_OSVs/OSV/UTC',\
+               'orbitX'   : './/Data_Block/List_of_OSVs/OSV/X',\
+               'orbitY'   : './/Data_Block/List_of_OSVs/OSV/Y',\
+               'orbitZ'   : './/Data_Block/List_of_OSVs/OSV/Z',\
+               }
+
+    # temp variables and parameters
+    container     = {}
+    # containerTemp = {}
+    events        = ('end',)
+
+    for key in queryList.keys():
+
+        try:
+            vars()[key]
+        except KeyError or NameError:
+           vars()[key] = []
+
+        for nodes in inTree.findall(queryList[key]):
+            vars()[key].append(nodes.text)
+
+        container[key] = vars()[key]
+    return container
+
+#--------------------------------------------------------
+def interpolate_orbit(input_orbit_dir, date, input_orbit_type, input_interpolation_method, satellite='S1A'):
+
+    orbit_time = calendar.timegm(time.strptime(date,'%Y-%m-%dT%H:%M:%S.%f'))
+    date_start = calendar.timegm(time.strptime(date[:10],'%Y-%m-%d'))
+
+    if input_orbit_type == 'POE':
+        input_orbit_dir = os.path.join(input_orbit_dir, 'precise')
+    elif input_orbit_type == 'RES':
+        input_orbit_dir = os.path.join(input_orbit_dir, 'restituted')
+
+    L = os.listdir(input_orbit_dir)
+    Orbit_info = []
+
+    if input_orbit_type == 'POE' and satellite == 'S1A':
+        orbit_type = 'S1A_OPER_AUX_POEORB_OPOD_'
+    elif input_orbit_type == 'RES' and satellite == 'S1A':
+        orbit_type = 'S1A_OPER_AUX_RESORB_OPOD_'
+    elif input_orbit_type == 'POE' and satellite == 'S1B':
+        orbit_type = 'S1B_OPER_AUX_POEORB_OPOD_'
+    elif input_orbit_type == 'RES' and satellite == 'S1B':
+        orbit_type = 'S1B_OPER_AUX_RESORB_OPOD_'
+
+    for d in L:
+        if d.startswith(orbit_type):
+            start_time = calendar.timegm(time.strptime(d[42:57], '%Y%m%dT%H%M%S'))
[…]
