+++ alignsar/itc-doris_5_patch2023/doris/prepare_stack/download_sentinel_data_orbits.py
@@ -1,0 +1,586 @@
+# This file contains a function to check which files for sentinel are available, which ones are downloaded and a quality
+# check for the files which are downloaded.
+
+import urllib
+import urllib2, urllib3
+import ssl
+import re
+import os, sys
+import datetime
+import base64
+import subprocess
+from fiona import collection
+from fastkml import kml
+from lxml import etree
+import xml.etree.ElementTree as ET
+
+
+def sentinel_available(start_day='', end_day='', sensor_mode='', product='', level='', track='', polarisation='', orbit_direction='', ROI='', user='', password=''):
+    # All available sentinel 1 images are detected and printed on screen.
+    # Following variables can be used to make a selection.
+    # shape > defining shape file or .kml
+    # start_day > first day for downloads (default one month before now) [yyyymmdd]
+    # end_day > last day for downloads (default today)
+    # track > the tracks we want to check (default all)
+    # polarisation > which polarisation will be used. (default all)
+
+    # string is the field we enter as url
+    string = ''
+
+    if sensor_mode:
+        string = string + ' AND ' + 'sensoroperationalmode:' + sensor_mode
+    if product:
+        string = string + ' AND ' + 'producttype:' + product
+    if level:
+        string = string + ' AND ' + level
+    if orbit_direction:
+        string = string + ' AND ' + 'orbitdirection:' + orbit_direction
+    if track:
+        string = string + ' AND ' + 'relativeorbitnumber:' + track
+    if start_day:
+        start = datetime.datetime.strptime(start_day, '%Y-%m-%d').strftime('%Y-%m-%d')
+    else:
+        start = (datetime.datetime.now() - datetime.timedelta(days=350)).strftime('%Y-%m-%d')
+    if end_day:
+        end = datetime.datetime.strptime(end_day, '%Y-%m-%d').strftime('%Y-%m-%d')
+    else:
+        end = datetime.datetime.now().strftime('%Y-%m-%d')
+    if polarisation:
+        string = string + ' AND ' + 'polarisationmode:' + polarisation
+    if ROI:
+        shape_str = load_shape_info(ROI)
+        string = string + ' AND footprint:"Intersects(POLYGON(' + shape_str + '))"'
+
+    date_string = 'beginPosition:[' + start + 'T00:00:00.000Z TO ' + end + 'T23:59:59.999Z] AND endPosition:[' + start + 'T00:00:00.000Z TO ' + end + 'T23:59:59.999Z]'
+    string = string + ' AND ' + date_string
+
+    # Finally we do the query to get the search result.
+    string = string[5:] + '&rows=1000'
+    url = 'https://scihub.copernicus.eu/dhus/search?q=' + urllib.quote_plus(string)
+    #url = 'https://scihub.copernicus.eu/gnss/search?q=' + urllib.quote_plus(string)# check for new orbit data downloading
+    #print(url)
+
+    print('Requesting available products: ' + url)
+    request = urllib2.Request(url)
+    base64string = base64.b64encode('%s:%s' % (user, password))
+    request.add_header("Authorization", "Basic %s" % base64string)
+
+    # connect to server. Hopefully this works at once
+    gcontext = ssl.SSLContext(ssl.PROTOCOL_TLSv1_2)
+    try:
+        dat = urllib2.urlopen(request, context=gcontext)
+    except:
+        print('not possible to connect this time')
+        return [], [], []
+
+    html_dat = ''
+    for line in dat:
+        html_dat = html_dat + line
+
+    parser = etree.HTMLParser()
+    tree = etree.fromstring(html_dat, parser)
+    products = [data for data in tree.iter(tag='entry')]
+    links = [data.find('link').attrib for data in tree.iter(tag='entry')]
+    dates = [data.findall('date')[1].text for data in tree.iter(tag='entry')]
+
[…]
