### 2025-01-02 a54d6b2
Delete itc-doris_5_patch2023 directory
