+++ itc-doris_5_patch2023/doris/envisat_tools/epr_api-2.2/src/epr_core.c
@@ -1,481 +1,0 @@
-/*
- * $Id: epr_core.c,v 1.1.1.1 2004-10-28 19:22:22 norman Exp $
- *
- * Copyright (C) 2002 by Brockmann Consult (info@brockmann-consult.de)
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the
- * Free Software Foundation. This program is distributed in the hope it will
- * be useful, but WITHOUT ANY WARRANTY; without even the implied warranty
- * of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
- */
-
-#include <assert.h>
-#include <errno.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-#include "epr_api.h"
-#include "epr_core.h"
-#include "epr_string.h"
-#include "epr_ptrarray.h"
-#include "epr_swap.h"
-#include "epr_field.h"
-#include "epr_record.h"
-#include "epr_param.h"
-#include "epr_dsd.h"
-#include "epr_msph.h"
-#include "epr_band.h"
-#include "epr_bitmask.h"
-
-#include "epr_dddb.h"
-
-
-/**
- * The one and only ENVISAT API instance.
- */
-EPR_SAPI epr_api;
-
-/*
-   Function:    epr_str_to_data_type_id
-   Access:      private API implementation helper
-   Changelog:   2002/01/05  nf initial version
- */
-/**
- * Converts the given string into a data type identifier.
- *
- * @param str the string to be converted.
- * @return the data type identifier represented by the given string.
- *         If the string can not be converted the special value
- *         <code>e_tid_unknown</code> is returned.
- */
-EPR_EDataTypeId epr_str_to_data_type_id(const char* str)
-{
-    assert(str != NULL);
-    if (epr_equal_names(str, "UChar") || epr_equal_names(str, "uchar"))
-        return e_tid_uchar;
-    else if (epr_equal_names(str, "AChar") || epr_equal_names(str, "SChar") || epr_equal_names(str, "char"))
-        return e_tid_char;
-    else if (epr_equal_names(str, "UShort") || epr_equal_names(str, "ushort"))
-        return e_tid_ushort;
-    else if (epr_equal_names(str, "SShort") || epr_equal_names(str, "short"))
-        return e_tid_short;
-    else if (epr_equal_names(str, "UInt") || epr_equal_names(str, "uint"))
-        return e_tid_uint;
-    else if (epr_equal_names(str, "SInt") || epr_equal_names(str, "int"))
-        return e_tid_int;
-    else if (epr_equal_names(str, "ULong") || epr_equal_names(str, "ulong"))
-        return e_tid_uint;
-    else if (epr_equal_names(str, "SLong") || epr_equal_names(str, "long"))
-        return e_tid_int;
-    else if (epr_equal_names(str, "Float") || epr_equal_names(str, "float"))
-        return e_tid_float;
-    else if (epr_equal_names(str, "Double") || epr_equal_names(str, "double"))
-        return e_tid_double;
-    else if (epr_equal_names(str, "@/types/UTC.dd") || epr_equal_names(str, "time"))
-        return e_tid_time;
-    else if (epr_equal_names(str, "String") || epr_equal_names(str, "string"))
-        return e_tid_string;
-    else if (epr_equal_names(str, "Spare") || epr_equal_names(str, "spare"))
[…]
