+++ alignsar/MLscripts/India - Object Detection/Data preparation/preprocessing.py
@@ -1,0 +1,266 @@
+import rioxarray
+import matplotlib.pyplot as plt
+import xarray as xr
+import numpy as np
+import matplotlib.cm as cm
+import netCDF4 as nc
+import os
+import cv2
+import  xml.dom.minidom
+import math
+import xmltodict
+import matplotlib.patches as patches
+import ast
+from collections import namedtuple
+from PIL import Image
+import joblib
+from joblib import Parallel, delayed
+import albumentations as A
+from matplotlib.image import imsave
+ 
+#Data preparation
+class Data_preparation:
+    """This class corresponds to the data preparation and subtiling"""
+    def read_data(path):
+        """This function reads and cleans the data"""
+        
+        # Read netCDF file
+        dataset = xr.open_dataset(path)
+        
+        #Convert dataset to pandas
+        data = dataset.to_dataframe()
+        
+        #Obtain entropy
+        entropy =dataset.variables['Entropy'].values
+        
+        #Remove nan
+        #Find indices that you need to replace
+        inds = np.isnan(entropy)
+        entropy[inds] = np.nanmedian(entropy)
+        #Place column means in the indices. Align the arrays using take
+        
+        #Read mask
+        l1_y =dataset.variables['Oil region'].values
+        
+        return entropy, l1_y
+    
+    def get_tiles(entropy,l1_y,data_folder):
+        """This function obtains the tiles for the image and the mask"""
+        
+        #Image dimensions
+        x0 = 4300
+        y0 = 2350
+        
+        #Get image tiles
+        l1_x_00 = entropy[:x0,:y0]
+        l1_x_01 = entropy[:x0,y0:]
+        l1_x_10 = entropy[x0:,:y0]
+        l1_x_11 = entropy[x0:,y0:]
+        
+        
+        #Get masked tiles
+        l1_y_00 = l1_y[:x0,:y0]
+        l1_y_01 = l1_y[:x0,y0:]
+        l1_y_10 = l1_y[x0:,:y0]
+        l1_y_11 = l1_y[x0:,y0:]
+        
+        #Save masks
+        plt.imsave(data_folder+"x_0.png", l1_x_00, cmap=cm.gray)
+        plt.imsave(data_folder+"x_1.png", l1_x_01, cmap=cm.gray)
+        plt.imsave(data_folder+"x_3.png", l1_x_10, cmap=cm.gray)
+        plt.imsave(data_folder+"x_4.png", l1_x_11, cmap=cm.gray)
+
+        plt.imsave(data_folder+"y_0.png", l1_y_00, cmap=cm.gray)
+        plt.imsave(data_folder+"y_1.png", l1_y_01, cmap=cm.gray)
+        plt.imsave(data_folder+"y_3.png", l1_y_10, cmap=cm.gray)
+        plt.imsave(data_folder+"y_4.png", l1_y_11, cmap=cm.gray)
+        
+        #Save images
+        plt.imsave(data_folder+"x.png", entropy, cmap=cm.gray)
+        plt.imsave(data_folder+"y.png", l1_y, cmap=cm.gray)
+    
+    
+class Augmentations:
+    """This class performs the different data augmentations to obtain a larger dataset"""
+    def pascalVOC_yolo(labels,labels_yolo):
+        """This function converts the original labels from pascalVOC xml file to yolo"""
+        
+        #Open Xml file
+        with open(labels, 'r') as file:
+            xml = file.read()
+
+        xml_dict = xmltodict.parse(xml)
+        class_dict = {'oil': 0}
+        
+        #Get image dimensions
+        img_w = int(xml_dict['annotation']['size']['width'])
+        img_h = int(xml_dict['annotation']['size']['height'])
+        
[…]

[…]
