+++ alignsar/MLscripts/readme.md
@@ -2,4 +2,5 @@ Three machine learning scripts are offered in this folder:
  - India - Object Detection: Yolov8 based oil spill detection in India
  - Netherlands - LULC: ANN based land use and land cover classification in the Netherlands
  - Poland - Change Detection: Siamese based change detection in Poland
-The test over the Netherlands, along with its sample data, is provided in 'tests' folder
+   
+The test over the Netherlands, along with its sample data, is provided in 'tests' folder (https://github.com/AlignSAR/alignSAR/tree/main/tests)
