+++ alignsar/MLscripts/readme.md
@@ -1,6 +1,6 @@
 Three machine learning scripts are offered in this folder: 
  - India - Object Detection: Yolov8 based oil spill detection in India
- - Netherlands - LULC: ANN based land use and land cover classification in the Netherlands
+ - Netherlands - LULC: ANN (artificial neural network) based land use and land cover classification in the Netherlands
  - Poland - Change Detection: Siamese based change detection in Poland
    
 The test over the Netherlands, along with its sample data, is provided in 'tests' folder (https://github.com/AlignSAR/alignSAR/tree/main/tests)
