+++ paper/paper.md
@@ -61,25 +61,25 @@ AlignSAR is available as an Apache-2.0-licensed open-source toolbox hosted on [G
 
 # Main Functionalities and Processing Flowchart
 
-1. [itc-doris_5_patch2023](https://github.com/AlignSAR/alignSAR/tree/main/itc-doris_5_patch2023): an updated version of Doris that is used to read raw SAR (Single Look Complex format), coregister all SAR images to the same reference grids, resample them, remove the reference and topographic phase using external Digital Elevation Model (DEM) (e.g [SRTM](https://www.earthdata.nasa.gov/sensors/srtm#:~:text=The%20Shuttle%20Radar%20Topography%20Mission,global%20dataset%20of%20land%20elevations.) or [Copernicus DEM](https://spacedata.copernicus.eu/collections/copernicus-digital-elevation-model)), and generate co-registered amplitude, SLC phase and interferometric phase data.
+1. [itc-doris_5_patch2023](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/itc-doris_5_patch2023): an updated version of Doris that is used to read raw SAR (Single Look Complex format), coregister all SAR images to the same reference grids, resample them, remove the reference and topographic phase using external Digital Elevation Model (DEM) (e.g [SRTM](https://www.earthdata.nasa.gov/sensors/srtm#:~:text=The%20Shuttle%20Radar%20Topography%20Mission,global%20dataset%20of%20land%20elevations.) or [Copernicus DEM](https://spacedata.copernicus.eu/collections/copernicus-digital-elevation-model)), and generate co-registered amplitude, SLC phase and interferometric phase data.
 
-2. [jupyter_notebook_demo](https://github.com/AlignSAR/alignSAR/tree/main/jupyter_notebook_demo): a jupyter notebook to demonstrate how to extract, visualize and analyse SAR signatures, with the sample output generated by Doris.
+2. [jupyter_notebook_demo](https://github.com/AlignSAR/alignSAR/tree/main/examples/jupyter_notebook_demo): a jupyter notebook to demonstrate how to extract, visualize and analyse SAR signatures, with the sample output generated by Doris.
 
-3. [rdcode](https://github.com/AlignSAR/alignSAR/tree/main/rdrcode): the radarcoding scripts to align other geospatial data onto SAR reference grid. For method details refer to @Kulshresta2023.
+3. [rdcode](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/rdrcode): the radarcoding scripts to align other geospatial data onto SAR reference grid. For method details refer to @Kulshresta2023.
 
-4. [snap_graphs](https://github.com/AlignSAR/alignSAR/tree/main/snap_graphs): graphs for SNAP used within the toolbox (e.g. rdr<->geocode).
+4. [snap_graphs](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/snap_graphs): graphs for SNAP used within the toolbox (e.g. rdr<->geocode).
 
-5. [stac](https://github.com/AlignSAR/alignSAR/tree/main/stac): python scripts to create STAC (SpatioTemporal Asset Catalogs). Our STAC examples are available via [STAC Browser](https://radiantearth.github.io/stac-browser/#/?.language=en) by specifying the STAC Catalog with 'https://data.crib.utwente.nl/stac/alignsar/catalog.json'.
+5. [stac](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/stac): python scripts to create STAC (SpatioTemporal Asset Catalogs). Our STAC examples are available via [STAC Browser](https://radiantearth.github.io/stac-browser/#/?.language=en) by specifying the STAC Catalog with 'https://data.crib.utwente.nl/stac/alignsar/catalog.json'.
    
[…]
