### 2025-01-05 AlignSAR/alignSAR#2 @LC-SAR
thank you so much @philippemiron for your valuable suggestions.
1. We have restructured the organization of the AlignSAR repo (https://github.com/AlignSAR/alignSAR) based on your suggestion.
2. We have provided the test (to validate the scripts and utilities) in the folders 'examples' (https://github.com/AlignSAR/alignSAR/tree/main/examples) and 'tests' (https://github.com/AlignSAR/alignSAR/tree/main/tests), along with brief instruction (see the files readme / README under a selected folder). We have provided all the information on AlignSAR installation, methods, data used and step-by-step procedure for the use case/example in the tutorial ‘AlignSAR_tutorial.pdf’ (https://github.com/AlignSAR/alignSAR/tree/main/tutorial).
3. Line 31-39 (https://github.com/AlignSAR/alignSAR/blob/main/paper/paper.md), we added the references for doris, snap, gamma, asf services - on demand processing and LiCSAR
4. We did add __pycache__ to .gitignore and remove it
5. We provided dockerfile installation in the ‘user_guide’ folder (https://github.com/AlignSAR/alignSAR/tree/main/paper/user_guide) and ‘README’ file (https://github.com/AlignSAR/alignSAR/blob/main/README.md), for the Dockerfile (https://github.com/AlignSAR/alignSAR/blob/main/Dockerfile). For instance, to build the dockerfile, run 'docker build -t alignsar .' and to test it, run 'docker run -it -v /your_local_path:/path_you_want_to_set_in_docker alignsar'. This dockerfile, as an example and demonstration, can automatically install Doris and relevant software.
6. This bashrc_alignsar is associated with the docker file 'DockerFile_LiCSAR_SNAP' (https://github.com/AlignSAR/alignSAR/blob/main/DockerFile_LiCSAR_SNAP). This dockerfile can install LiCSAR and SNAP software. The methods and workflow description can be found in the tutorial ‘AlignSAR_tutorial.pdf’. In case of not building and running the Dockerfile(s), one can also manually install Doris, LiCSAR and SNAP and customize preferred processing environment.
7.  We double-checked the two dockerfiles. For the DockerFile you ran, Line 23 (https://github.com/comet-licsar/licsar_extra.git) does not work, and we are working on it. As licsar_extra.git is not used for the AlignSAR version 1.1 (submitted for JOSS). Please test https://github.com/AlignSAR/alignSAR/blob/main/Dockerfile, by following the instruction we provided.

Thanks again and looking foward to your further suggestions, with regards,

Ling
