### 2025-01-05 AlignSAR/alignSAR#3 @LC-SAR
Hi @shahchiragh, thanks so much for your suggestions.
1. Codebase Organization: we have restructured the organization of the AlignSAR repo (https://github.com/AlignSAR/alignSAR) based on your suggestion.
2. Unwanted Files/Directories: while restructuring the organization of our repo, we have removed unwanted files and directories.
3. Dockerfile Naming: we double-checked the two dockerfiles. For the DockerFile you ran, Line 23 (https://github.com/comet-licsar/licsar_extra.git) does not work, and we are working on it. As licsar_extra.git is not used for the AlignSAR version 1.1 (submitted for JOSS) and simplify the test, please test https://github.com/AlignSAR/alignSAR/blob/main/Dockerfile, by following the instruction we provided. This 'https://github.com/AlignSAR/alignSAR/blob/main/Dockerfile' automatically installs Doris and relevant software. We provided the dockerfile installation in the ‘user_guide’ folder (https://github.com/AlignSAR/alignSAR/tree/main/paper/user_guide) and ‘README’ file (https://github.com/AlignSAR/alignSAR/blob/main/README.md), for the Dockerfile (https://github.com/AlignSAR/alignSAR/blob/main/Dockerfile). For instance, to build the dockerfile, run 'docker build -t alignsar .' and to test it, run 'docker run -it -v /your_local_path:/path_you_want_to_set_in_docker alignsar'.
4. Alternate Code Installation Process: in the tutorial (https://github.com/AlignSAR/alignSAR/blob/main/tutorial/AlignSAR_tutorial.pdf), Appendix A (page 30), we describe how to install e.g. Doris and relevant software manually.
5. Tests for Validation: We have provided the test (to validate the scripts and utilities) in the folders 'examples' (https://github.com/AlignSAR/alignSAR/tree/main/examples) and 'tests' (https://github.com/AlignSAR/alignSAR/tree/main/tests), along with brief instruction (see the files readme / README under a selected folder). We have provided all the information on AlignSAR installation, methods, data used and step-by-step procedure for the use case/example in the tutorial ‘AlignSAR_tutorial.pdf’ (https://github.com/AlignSAR/alignSAR/tree/main/tutorial).

Thanks again and looking forward to your further suggestions.

Regards,

Ling
