### 2025-01-08 AlignSAR/alignSAR#2 @philippemiron
Thanks for those updates. I have a few more suggestions to be able to close the last points.

- [ ] Installation instructions. Initially, building the Docker image failed, even specifying the platform, `docker build—-platform=linux/amd64 -t alignsar`. I get the following error:
```
206.4   Could not find a version that satisfies the requirement fiona (from versions: )
206.4 No matching distribution found for fiona
------
Dockerfile:12
--------------------
  11 |     # Build Python source for doirs
  12 | >>> RUN apt-get install -y gdal-bin python-gdal libgdal-dev gawk make tcsh csh wget git tar zip vim build-essential python-pip \
  13 | >>>   && pip install fiona shapely scipy html HTMLParser requests fastkml lxml -U pyopenssl matplotlib pyproj
  14 |
```
- [ ] How is the user supposed to install Doris-5 installation which seems to be required to run the Docker image?
- [ ] Example usage and Automated tests. It would be ideal to include an `pyproject.toml`, or a conda `environment.yml in the example folders to be able to install quickly the required packages to run the example. Currently, this is left to the user to figure out. This also prevent me from validating the tests.
