+++ pyproject.toml
@@ -1,0 +1,43 @@
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "alignsar"
+version = "1.1"
+requires-python = ">=3.8"
+dependencies = [
+    "numpy",
+    "scikit-learn",
+    "scipy",
+    "xarray",
+“netCDF4”,
+“matplotlib”,
+“resdata”,
+“rasterio”,
+“geopandas”,
+“pyproj”,
+“gdal”,
+“rioxarray”
+]
+description = "AlignSAR: An open-source toolbox of SAR benchmark dataset creation for machine learning applications"
+readme = "README.md"
+license = {file = "LICENSE.txt"}
+authors = [
+    {name = "Ling Chang", email = "ling.chang@utwente.nl"},
+    {name = "Xu Zhang"},
+    {name = "Anurag Kulshrestha"},
+    {name = "Milan Lazecky"},
+    {name = "Serkan Girgin"},
+    {name = "Jose Manuel Delgado Blasco"},
+]
+keywords = ["SAR", "Benchmark Datasets", "Machine Learning"]
+classifiers=[
+    'Intended Audience :: Science/Research',
+    'License :: OSI Approved :: Apache Software License',
+    'Natural Language :: English',
+    'Programming Language :: Python :: 2.7.17',
+    'Programming Language :: Python :: 3.10']
+
+[project.urls]
+repository = " https://github.com/AlignSAR/alignSAR"
