### 2025-01-14 8c311fb
Delete pyproject.toml
