### 2025-01-14 AlignSAR/alignSAR#2 @LC-SAR
Hi @philippemiron, thanks for your follow-up comments. Please see the response below:

1. Installation instructions: We have tested our dockerfile with three different operating systems/docker versions: macOS (docker version: version 4.37.2), ubuntu 20.04.6 LTS (docker version: 24.0.5) and Ubuntu 20.04.1 (docker version: 24.0.7), we didn’t encounter any issues when building/running dockerfile.
For instance, with macOS (Ventura 13.7.2), docker-desktop app (version 4.37.2(179585)), here are the test log, see the screenshot below.
![Screenshot 2025-01-13 at 14 37 36](https://github.com/user-attachments/assets/bf0ec211-33b8-4032-8aed-3f50317e5190)
We didn’t figure out why you had this error and how to solve it.
 2. How is the user supposed to install Doris-5 installation: We have addressed that ‘_Note that AlignSAR provides SAR data preprocessing workflow using 1) Doris, and 2) LiCSAR and SNAP. ‘Dockerfile’ can install Doris-5 and its relevant third-party software tools, while DockerFile_LiCSAR_SNAP is another dockerfile we offered, which can install LiCSAR and SNAP software. **The methods and workflow description can be found in the tutorial ‘AlignSAR_tutorial.pdf’. In case of not building and running the Dockerfile(s), one can also manually install Doris, LiCSAR and SNAP and customize preferred processing environment.**_’ in https://github.com/AlignSAR/alignSAR/blob/main/README.md. And in the tutorial https://github.com/AlignSAR/alignSAR/blob/main/tutorial/AlignSAR_tutorial.pdf, Appendix A (page 30) describes how to manually install Doris.
 3. Example usage and Automated tests: We created pyproject.toml file, see https://github.com/AlignSAR/alignSAR/blob/main/pyproject.toml. Note that Doris-5 is built upon Python version-2.7.17, which is used to generate SAR signatures (such as interferograms and amplitude maps). The procedure is illustrated in the tutorial https://github.com/AlignSAR/alignSAR/blob/main/tutorial/AlignSAR_tutorial.pdf, Chapter 3 ‘Procedure description’. In https://github.com/AlignSAR/alignSAR/blob/main/examples/data/data_links.txt, for instance, we shared the samples of the resultant SAR benchmark datasets (using Doris) for the use case in the Netherlands (item 5). You can test them with jupyter_notebook_demo (https://github.com/AlignSAR/alignSAR/tree/main/examples/jupyter_notebook_demo). The rest of the tools we created are built upon Python version 3.10, therefore in project.toml, we list two Python language versions: 2.7.17 and 3.10.
