### 2025-01-20 ae6f745
Update batch_LiCSBAS.sh
