### 2025-01-20 4d2cfd8
Update batch_LiCSBAS.sh
