### 2025-01-23 dd56934
Update batch_LiCSBAS.sh
