### 2025-01-24 425e102
Create run_licsalert.py
