### 2025-02-04 AlignSAR/alignSAR#3 @shahchiragh
Hi @LC-SAR,

Thanks for making the necessary updates! The changes look good to me so far. However, I'm still running into some issues with the installation step. I'm not sure if there's an additional step I'm missing or if there's an opportunity to simplify the process for a smoother installation. This might be related to the Python version differences.

Here is a snapshot of the latest error I'm encountering:

```
361.1 Get:293 http://ports.ubuntu.com/ubuntu-ports bionic/universe arm64 proj-bin arm64 4.9.3-2 [30.8 kB]
361.2 E: Failed to fetch http://ports.ubuntu.com/ubuntu-ports/pool/main/libs/libsigsegv/libsigsegv2_2.12-1_arm64.deb  Undetermined Error [IP: 185.125.190.39 80]
361.2 E: Unable to fetch some archives, maybe run apt-get update or try with --fix-missing?
361.2 Fetched 154 MB in 6min 0s (426 kB/s)
------
Dockerfile:12
--------------------
  11 |     # Build Python source for doirs
  12 | >>> RUN apt-get install -y gdal-bin python-gdal libgdal-dev gawk make tcsh csh wget git tar zip vim build-essential python-pip \
  13 | >>>   && pip install fiona shapely scipy html HTMLParser requests fastkml lxml -U pyopenssl matplotlib pyproj
  14 |
--------------------
ERROR: failed to solve: process "/bin/sh -c apt-get install -y gdal-bin python-gdal libgdal-dev gawk make tcsh csh wget git tar zip vim build-essential python-pip   && pip install fiona shapely scipy html HTMLParser requests fastkml lxml -U pyopenssl matplotlib pyproj" did not complete successfully: exit code: 100
```
