### 2025-02-04 AlignSAR/alignSAR#3 @LC-SAR
@shahchiragh thanks so much for your follow-up. The reviewer 'philippemiron' raised a similar issue, please see our answer on https://github.com/AlignSAR/alignSAR/issues/2#issuecomment-2590728726
It could be due to different versions of Python. For this project, we used two Python versions. As Doris is designed upon old version of Python (2.7.17), the docker pre-installs this old version of Python. The post-analysis tool we developed, is built up Python 3.10
If the error persists, should we arrange a brief online meeting? [I'm not sure if this is permitted for JOSS reviewing.]
