### 2025-02-05 #7532 @rwegener2
Hi there everyone,

In reading over the status of this paper review I'm noticing some themes from both reviewers. It seems that installation and the ability to find and run tests have been difficult for both reviewers. It appears that some of this information is available, but was perhaps hard to find. @philippemiron and @shahchiragh, if either of you have suggestions for how the authors could reorganize to make the tests easier to run or locate those could be useful pieces of feedback for the authors.

As for installation, that is an important point for the authors to address. It also hinders the ability of a reviewer to run evaluate the testing and functionality aspects of the review. @LC-SAR if there is any information (Ex. operating systems and versions) that the reviewers can provide to help you debug these installation problems please feel free to ask follow up questions.
