### 2025-02-06 #7532 @LC-SAR
Thank you @rwegener2 for your comments and suggestions. We have updated the repo based on two reviewers' suggestions. For theirlast comments on docker installation, we tested internally again with three different operating systems/software, all work from our side. We await two reviewers' further responses, and then we will update the information related to e.g. operating systems and versions on the github repo. Thank you
