### 2025-02-07 AlignSAR/alignSAR#4 @rwegener2
Installation Instructions Docs
Thanks for providing installation instructions with this software. It's a great first step for a user. After installing the software I have a few suggestions on the installation instructions:

- [ ] There appear to be several different options for how to install this software and I got lost trying to understand what those options were. Example 1: "The user can either use Docker with our provided Docker file, or download SNAP via ESA SNAP, and Doris-5 via GitHub AlignSAR and manually install them on the computer." -> Are there three options here (1. dockerfile 2. snap, 3. doris-5 via github) or two (1. dockerfile 2. snap and doris 5)? Example 2: "Another solution is to use another DockerFile stored in the folder ’itc-doris 5 patch2023’ with the same name DockerFile." In what scenario would I need to do this? If I used the original Dockerfile or if I didn't? I would suggest adding sub-headings to the installation to help guide users through the steps and the options at each step.

For example (feel free to rearrange as you see fit to follow the correct flow in installation)
Installation
1. Set up environment
a. Option 1: Use dockerfile
b. Option 2: Manually install SNAP and Doris-5
2. Setting up SNAP and Doris-5
a. Option 1: ...
4. Mounting the AlignSAR repository

- [ ] The link to the Doris-5 "itc-doris 5 patch2023" goes to a webpage with a 404. I would also suggest writing a filepath to the download file in the documentation.

- [ ] What does "The instruction of Docker software installation on e.g. Ubuntu refers to dokcer-docs,
and DigitalOcean." mean? Is is that the instructions for how to install Docker on Ubuntu are available via Digital Ocean?

- [ ] "Here ’alignsar’ is to be mounted, and can be customized by the enduser." I would provide one example syntax for this, in case the user is unfamiliar with how to mount a volume.

related to openjournals/joss-reviews#7532
