### 2025-02-07 #7532 @rwegener2
Hi @LC-SAR. I understand that it can be difficult to test and account for every possible architecture. Still, it is important to remember that our reviewers are volunteers and debugging an installation can be quite time intensive, especially when they are new to a software. I also went through and tested the installation. It was successful when I installed on my Mac (Sonoma 14.6.1). I provided some suggestions for improving the installation instructions in a repository issue.

Hopefully between myself and the reviewers we can help ensure the AlignSAR installation and the instructions are as clear and accessible to as many people as possible. It seems like perhaps the Mac installation is operating well, but @philippemiron and @shahchirag are operating on Linux systems (is that correct?). Can you both please provide your linux flavors and versions?

I see a few possible routes forward:
1. @LC-SAR from [this comment](https://github.com/AlignSAR/alignSAR/issues/2#issuecomment-2590728726) it sounds like you already have a system for testing multiple infrastructures. You could try the installation using the same linux versions as the reviewers in hopes of reproducing their errors.
2. It is certainly fair if you are not interested in supporting every operating system/version. Another option would be to include in your documentation that only certain operating systems and linux versions are supported. In that case I could begin a search for another set of reviewers that are able to meet the operating system requirements of this library.

I think this is a great software for the community and I'd love to see it keep moving forward. Let me know what you think the best next step is.
