### 2025-02-08 AlignSAR/alignSAR#3 @shahchiragh
Hi @LC-SAR -- No problem at all, and I appreciate your patience with the installation issues. I understand that the process can vary based on several factors. However, I was able to dedicate some time to thoroughly troubleshoot the installation steps, and I identified the necessary modifications to the Dockerfile that successfully resolved the issues.

For reference, I am using macOS Sequoia 15.2 and had to adjust my Python version to 2.7.18. Despite this change, I still encountered multiple installation errors. However, after several iterations, the following Dockerfile configuration ultimately worked. It might help, if you would like to involve this as troubleshooting for different OS or setups.

```
FROM ubuntu:18.04
USER root
# Install dependencies
# PIP - openssl version > 1.1 may be an issue (try older ubuntu images)
RUN apt-get update \
  && apt-get install -y wget gcc make openssl libffi-dev libgdbm-dev libsqlite3-dev libssl-dev zlib1g-dev \
  && apt-get install -y python-pip python-dev python-setuptools \
  && apt-get install -y gfortran libopenblas-dev liblapack-dev \
  && apt-get clean

# Build Python source for doirs
RUN apt-get install -y gdal-bin python-gdal libgdal-dev gawk make tcsh csh wget git tar zip vim build-essential python-pip \
    libxml2 \
    libxml2-dev \
    libxslt1-dev \
    zlib1g-dev \
    libfreetype6-dev \
    pkg-config \
    proj-bin \
    libproj-dev

# Upgrade pip, setuptools
RUN pip install --upgrade pip setuptools

# Install specific versions of cffi and cryptography
RUN pip install cffi==1.12.3
RUN pip install cryptography==2.8

# Install Python dependencies
RUN pip install fiona shapely scipy html HTMLParser requests fastkml lxml pyopenssl matplotlib pyproj==1.9.6

# Install FFTW and Doris
RUN apt-get update \
  && wget -c http://www.fftw.org/fftw-3.3.10.tar.gz \
  && tar -xvf fftw-3.3.10.tar.gz \
  && cd fftw-3.3.10 \
  && ./configure --enable-float \
  && make \
  && make install \
  && cd ~ \
  && git clone https://github.com/LC-SAR/Doris5ITCupdate.git \
  && cd Doris5ITCupdate/doris/doris_core \
  && chmod +x ./configure \
  && ./configure \
  && make \
  && make install \
  && cd ../sar_tools \
  && make \
  && make install \
  && cd ../envisat_tools \
  && make \
  && make install
```

Lastly, I believe the post-installation steps may require a few path adjustments. Specifically, after building the image and starting the container, the path may need to be updated from /root/DorisITCupdate to /root/Doris5ITCupdate to ensure proper functionality.

```
Afterward, the Doris-5 installation directory should be modified by editing the files /root/DorisITCupdate/doris/doris_stack/main_code/doris_main.py (Line 5) and /root/DorisITCupdate/doris/install/doris_config.xml (Line 2). Specifically, Line 5: ‘sys.path.append(‘/home/username/software/’)’ should be changed to ‘sys.path.append(‘/root/Doris5ITCupdate/’)’ and Line 2: change to ‘<source_path>/root/Doris5ITCupdate/doris</source_path>’.
```

Everything else looks great! I'll move forward with the remaining items for the review. Thanks again, and best wishes!
