### 2025-02-08 #7532 @shahchiragh
Hi @rwegener2, apologies for the delay in getting back and troubleshooting this issue, as I understand it may have added extra work for the authors. I was able to take out some time to thoroughly troubleshoot the installation steps and identified the necessary modifications to the Dockerfile that resolved the issues. I’m using macOS Sequoia 15.2 and I had to adjust my Python version to 2.7.18. Even after the Python version change, I ran into several installation errors, but with some additional tweaks, I was able to get everything working. I've shared these changes with @LC-SAR  in [Issue 3](https://github.com/AlignSAR/alignSAR/issues/3) as a reference for troubleshooting, especially for users with different setups. I'll now proceed with the remaining items in the review, but if there's anything else you’d like me to check, please don’t hesitate to reach out.
