+++ README.md
@@ -2,7 +2,7 @@
 The AlignSAR tools are used to extract representative SAR signatures, and ultimately offer FAIR-guided open SAR benchmark dataset library designed for SAR-based artificial intelligence applications, while ensuring interoperability and consistency with existing and upcoming initiatives and technologies, facilitating wider exploitation of SAR data and its integration and combination with other datasets. This library will contain meaningful and accurate SAR signatures created by integrating and aligning multi-SAR images and other geodetic measurements in time and space. Related link: https://www.alignsar.nl
 
 ## Tool description:
-[Dockerfile](https://github.com/AlignSAR/alignSAR/blob/main/Dockerfile): docker file with OS and software setup. 'Dockerfile' is to install Doris-5 and relevant software, 'DockerFile_LiCSAR_SNAP' is to install LiCSAR and SNAP.
+Dockerfile: docker file with OS and software setup. ['Dockerfile'](https://github.com/AlignSAR/alignSAR/blob/main/Dockerfile) is to install Doris-5 and relevant software, 'DockerFile_LiCSAR_SNAP' is to install LiCSAR and SNAP.
 
  in 'alignsar' folder
 1. MLscripts: scripts for machine learning analysis. Yolov8, ANN and Siamese are separately used for Object Detection (India), Land Use Land Cover classification (Netherlands), and Change Detection (Poland).  
