### 2025-02-12 #7532 @LC-SAR
Dear @shahchiragh, thanks for taking the time to test and offer suggestions on improving our dockerfile. We tested the dockerfile you suggested on mac OS successfully, and included it in our github repo, see https://github.com/AlignSAR/alignSAR/blob/main/misc/Dockerfile1. In Dockerfile1, we only modified the version of fftw from 3.3.10 to 3.2.2, as the higher version of fftw is not compatible with Doris software. We also supplemented the instruction in https://github.com/AlignSAR/alignSAR/blob/main/README.md, explaining the '[Dockerfile](https://github.com/AlignSAR/alignSAR/blob/main/Dockerfile) works well with three different operating systems/docker versions: macOS Ventura 13.7.3 (docker version: version 4.37.2), ubuntu 20.04.6 LTS (docker version: 24.0.5) and Ubuntu 20.04.1 (docker version: 24.0.7). If you encounter version-related errors or missing Python packages while running this [Dockerfile](https://github.com/AlignSAR/alignSAR/blob/main/Dockerfile) on macOS Sequoia 15.2, please use [Dockerfile1](https://github.com/AlignSAR/alignSAR/blob/main/misc/Dockerfile1) in the folder 'misc' instead.'

Looking forward to your further suggestions.
