### 2025-03-14 #7532 @rwegener2
@philippemiron and @shahchiragh How are things looking on the review?
@philippemiron I would still interested in what operating system you are using. If you are unable to install that is alright. This software may not support all operating systems/versions so if doesn't work on your operating system that is good feedback for the documentation. In that case I will search for a third reviewer that meets the system requirements for this software and can give a full review. Any other thoughts you have on improving the documentation are still welcome, even if you can't install.

@shahchiragh do you think the new dockerfile is sufficient and well explained in the documentation? Do you have any feedback on the other items in the review checklist?

@LC-SAR I made some comments on the documentation in this issue https://github.com/AlignSAR/alignSAR/issues/4. Please take a moment to address those pieces of feedback as well. I think they will help your users better understand how to get setup.
