### 2025-03-18 #7532 @shahchiragh
@rwegener2 Thanks for the ping! The changes made by the author, including the new Dockerfile (https://github.com/AlignSAR/alignSAR/blob/main/misc/Dockerfile1) and related documentation (https://github.com/AlignSAR/alignSAR/blob/main/README.md), look good to go. I also agree with the comments on the issue (https://github.com/AlignSAR/alignSAR/issues/4), as they would address the `Quality of writing` review item.

@LC-SAR I don't see any references, are there not or they are missing in the paper md file?
