### 2025-03-19 #7532 @LC-SAR
@shahchiragh, thanks for your rapid response and further comment. Regarding the reference in paper.md (https://github.com/AlignSAR/alignSAR/blob/main/paper/paper.md), it is not shown in paper.md version, but it is shown in paper.pdf when running https://github.com/AlignSAR/alignSAR/blob/main/.github/workflows/draft_pdf.yml, see an outcome example: https://github.com/AlignSAR/alignSAR/actions/runs/12588232946. Here I also share the current version of paper.pdf via https://surfdrive.surf.nl/files/index.php/s/HMlRUAGUlpqO3BF for your review. from page 3 to 4, all the references are listed in paper.pdf. I think it is standard that the reference is not shown in paper.md, see JOSS examples https://github.com/VegeWaterDynamics/motrainer/blob/main/paper/paper.md and https://github.com/enthusi/feline/blob/main/joss/paper.md. Hope my answer is satisifed. regards.
