+++ alignsar_extension_InSAR_TS/README.md
@@ -2,6 +2,7 @@
 Milan Lazecky, University of Leeds
 
 Support materials to generate AlignSAR InSAR Time Series Signature Datacube (ITSSD):
+- 022D_ALIGNSAR.nc - demo datacube over Campi Flegrei data, as described in AlignSAR InSAR TS Extension MUC
 - data_coverage.kml - a Google Earth KML file containing coverage polygon of the sample dataset
 - input_S1_files.txt - input Sentinel-1 file names that were used to generate LiCSAR interferograms over the sample area of interest (AOI)
 - prelim_*tif - preliminary outputs serving as a preview of the content of the final datacubes
