### 2025-03-23 #7532 @philippemiron
Hi everyone, I was able to build the image correctly with the latest change. My main comments are that the library should be made easier to test and validate.

For example, would it be possible to run the tests either:
- automatically with GitHub actions;
- or in the docker images

Then, similarly for a real-work example. It would be ideal if you have a Notebook that downloads the required data, and uses alignSAR to do the processing. Right now, a lot of those steps are left to the users, and it makes the learning curves quite steep.
