### 2025-03-26 #7532 @rwegener2
Thanks for those updates @philippemiron and @LC-SAR.

@LC-SAR would you please provide a bit more background about the files you have added to the `tests` folder. They seem to be nice example ML scripts, but they don't appear to import any modules from the Align-SAR package. Does Align-SAR include an installable Python package, or is it's benefit in providing an environment and a series of scripts to assist users in SAR data processing?
