### 2025-03-27 #7532 @rwegener2
Hi @LC-SAR,

After reviewing all of the cumulative work by the reviewers and the state of the repository I am going to send this submission to the track editor in chief for a scope query. At this point I believe that substantial work is needed to bring this repository to a submission-level state. This includes:
- **Testing**: Tests in a processing toolbox should address the core functionality of the toolbox, not the outputs of the toolbox. In this case that means that tests need to generate the datasets, not use them. This will be a significant piece of work.
- **General organization**: Many tool submissions to JOSS are packaged as an installable programming library (ex. pip installable Python tool). This provides a level of abstraction which makes the functionality easier to execute as a user and also provides a predictable structure for developers to navigate. It is not a requirement that JOSS submissions be installable libraries, but a submission which does not follow this structure still needs to provide a similar level of clarity to the users and the developers. At this point the file structure, readme, and user resources are not organized in a way that is sufficiently clear to users or developers. This will require some work, either through restructuring the method of distribution or through dedicated effort to streamline and detail the readme, the steps for installation and script execution.
- **Installation instructions**: Multiple reviewers had a difficult time following the installation instructions, which are quite complicated. https://github.com/AlignSAR/alignSAR/issues/4 provides multiple specific ways to improve the installation instructions.

It would be my suggestion that the authors look at other ways to package this library, as that would provide developers and users more clarity. Following a standard for package installation would also supply a common framework for developing tests and creating a documentation page.

The track editor will take on the task of reviewing this submission and the reviewer comments thus far, but I would like to express that I hope this toolbox is re-submitted to JOSS. There is a good level of useful analysis in the package that, especially with improved level of distribution, provides great community benefit.
