### 2025-04-17 #7532 @rwegener2
Hi @LC-SAR, just checking in on the status of the updates. I imagine it is a lot of work to make the changes described above and it may take some time. I just want to re-iterate that if we close this submission you are free to re-submit to JOSS at any time. There won't be any penalty to having already submitted. In that case you can take as long as you want to make updates to the repository.
If you don't want to go that route and still like to make updates to this submission please do so by April 28th. We have many waitlisted submissions that would also like to receive reviewer feedback.

If you have any questions please feel free to ask.
