+++ README.md
@@ -5,18 +5,18 @@ The AlignSAR tools are used to extract representative SAR signatures, and ultima
 Dockerfile: docker file with OS and software setup. ['Dockerfile'](https://github.com/AlignSAR/alignSAR/blob/main/Dockerfile) is to install Doris-5 and relevant software, 'DockerFile_LiCSAR_SNAP' is to install LiCSAR and SNAP.
 
  in 'alignsar' folder
-1. MLscripts: scripts for machine learning analysis. Yolov8, ANN and Siamese are separately used for Object Detection (India), Land Use Land Cover classification (Netherlands), and Change Detection (Poland).  
-2. bin: non-python scripts used within the toolbox (e.g. rdr<->geocode)
-3. itc-doris_5_patch2023: an updated version of Doris-5 software developed by TUDelft.
-4. rdcode: the radarcoding scripts.
-5. snap_graphs: graphs for SNAP used within the toolbox (e.g. rdr<->geocode)
-6. stac: python scripts to create STAC
-7. Meta_info_extraction_global_local.py: script to extract global/local attributes from Sentinel-1 SAR metadata. 
-8. bashrc_alignsar: install settings for the expected environment variables and paths
-9. resdata.py: script needed from signature_extraction.py.
-10. signature_extraction.py: SAR signature extraction script.
-11. speckle_filt.py: A speckle filtering script.
-12. alignsar_utils.py: various python functions used within the toolbox.
+1. bin: non-python scripts used within the toolbox (e.g. rdr<->geocode)
+2. itc-doris_5_patch2023: an updated version of Doris-5 software developed by TUDelft.
+3. rdcode: the radarcoding scripts.
+4. snap_graphs: graphs for SNAP used within the toolbox (e.g. rdr<->geocode). [optional, having the same function as rdcode]
+5. stac: python scripts to create STAC
+6. Meta_info_extraction_global_local.py: script to extract global/local attributes from Sentinel-1 SAR metadata. 
+7. bashrc_alignsar: install settings for the expected environment variables and paths
+8. resdata.py: script needed from signature_extraction.py.
+9. signature_extraction.py: SAR signature extraction script.
+10. speckle_filt.py: A speckle filtering script.
+11. alignsar_utils.py: various python functions used within the toolbox.
+12. MLscripts: scripts for machine learning analysis. Yolov8, ANN and Siamese are separately used for Object Detection (India), Land Use Land Cover classification (Netherlands), and Change Detection (Poland). [for SAR benchmark dataset post-analysis] 
 
 in 'examples' folder
 1. jupyter_notebook_demo: a jupyter notebook to demonstrate how to extract, visualize and analyse SAR signatures.
