+++ paper/paper.md
@@ -65,9 +65,7 @@ AlignSAR is available as an Apache-2.0-licensed open-source toolbox hosted on [G
 
 2. [jupyter_notebook_demo](https://github.com/AlignSAR/alignSAR/tree/main/examples/jupyter_notebook_demo): a jupyter notebook to demonstrate how to extract, visualize and analyse SAR signatures, with the sample output generated by Doris.
 
-3. [rdcode](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/rdrcode): the radarcoding scripts to align other geospatial data onto SAR reference grid. For method details refer to @Kulshresta2023.
-
-4. [snap_graphs](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/snap_graphs): graphs for SNAP used within the toolbox (e.g. rdr<->geocode).
+3. [rdcode](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/rdrcode): the radarcoding scripts to align other geospatial data onto SAR reference grid. For method details refer to @Kulshresta2023. Another option is [snap_graphs](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/snap_graphs)that are graphs for SNAP used within the toolbox (e.g. rdr<->geocode).
 
 5. [stac](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/stac): python scripts to create STAC (SpatioTemporal Asset Catalogs). Our STAC examples are available via [STAC Browser](https://radiantearth.github.io/stac-browser/#/?.language=en) by specifying the STAC Catalog with 'https://data.crib.utwente.nl/stac/alignsar/catalog.json'.
    
@@ -77,9 +75,7 @@ AlignSAR is available as an Apache-2.0-licensed open-source toolbox hosted on [G
 
 8. [signature_extraction.py](https://github.com/AlignSAR/alignSAR/blob/main/alignsar/signature_extraction.py): a script to extract representative SAR signatures.
 
-9. [speckle_filt.py](https://github.com/AlignSAR/alignSAR/blob/main/alignsar/speckle_filt.py): a script for spatio-temoporal speckle filtering. For more details, refer to @Quegan2001.
-
-10. [MLscripts](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/MLscripts): scripts for machine learning analysis. [Yolov8](https://github.com/ultralytics/ultralytics), ANN (Artificial Neural Network) and Siamese are separately used for Object Detection (India), Land Use Land Cover classification (Netherlands), and Change Detection (Poland). 
+9. [MLscripts](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/MLscripts): scripts for machine learning analysis. [Yolov8](https://github.com/ultralytics/ultralytics), ANN (Artificial Neural Network) and Siamese are separately used for Object Detection (India), Land Use Land Cover classification (Netherlands), and Change Detection (Poland). 
 
 Figure [1](flowchart.png) illustrates the processing flowchart for creating SAR benchmark datasets, with key functionalities highlighted in blue. The flowchart consists of six main stages: environment and software setup, pre-processing, signature extraction, SAR benchmark dataset generation, STAC creation, and machine learning application demonstration.
 ![](flowchart.png) 
