+++ paper/paper.md
@@ -65,7 +65,7 @@ AlignSAR is available as an Apache-2.0-licensed open-source toolbox hosted on [G
 
 2. [jupyter_notebook_demo](https://github.com/AlignSAR/alignSAR/tree/main/examples/jupyter_notebook_demo): a jupyter notebook to demonstrate how to extract, visualize and analyse SAR signatures, with the sample output generated by Doris.
 
-3. [rdcode](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/rdrcode): the radarcoding scripts to align other geospatial data onto SAR reference grid. For method details refer to @Kulshresta2023. Another option is [snap_graphs](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/snap_graphs)that are graphs for SNAP used within the toolbox (e.g. rdr<->geocode).
+3. [rdcode](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/rdrcode): the radarcoding scripts to align other geospatial data onto SAR reference grid. For method details refer to @Kulshresta2023. Another option is [snap_graphs](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/snap_graphs) that are graphs for SNAP used within the toolbox (e.g. rdr<->geocode).
 
 5. [stac](https://github.com/AlignSAR/alignSAR/tree/main/alignsar/stac): python scripts to create STAC (SpatioTemporal Asset Catalogs). Our STAC examples are available via [STAC Browser](https://radiantearth.github.io/stac-browser/#/?.language=en) by specifying the STAC Catalog with 'https://data.crib.utwente.nl/stac/alignsar/catalog.json'.
    
