+++ tests/Meta_info_extraction_global_local.py
@@ -1,0 +1,205 @@
+import xml.etree.ElementTree as ET
+import os
+
+def get_xml_path(sar_folder_path, folder_num, xml_num):
+    items = os.listdir(sar_folder_path)
+    # create 'folders' list to contain all the sar image folders, e.g., folders=[S1A_IW_SLC__1SDV_20220109T171712_20220109T171740_041387_04EBB7_25EB.SAFE, ...]
+    folders = [item for item in items if os.path.isdir(os.path.join(sar_folder_path,item))]
+    # choose one sar image folder and come into it, attach xml folder 'annotation' to it
+    xml_path = sar_folder_path + folders[folder_num] + '/annotation'
+    items = os.listdir(xml_path)
+    # create 'files' list to contain all the xml files, e.g., files=[s1a-iw1-slc-vh-20220109t171712-20220109t171738-041387-04ebb7-001.xml, ...]
+    files = [item for item in items if os.path.isfile(os.path.join(xml_path,item))]
+    # return one of the xml file path, depend on 'xml_num'
+    return sar_folder_path + folders[folder_num] + '/annotation/' + files[xml_num]
+
+
+def get_global_attribute(sar_folder_path, folder_num, xml_num):
+    xml_path = get_xml_path(sar_folder_path,folder_num,xml_num)
+    tree = ET.parse(xml_path)
+    root = tree.getroot()
+
+    swath = root.find(".//swath").text
+
+    sar_date_time = root.find(".//startTime").text[:10] # end to date
+    sar_master_date_time = "20220214"
+    sar_UTC_time = root.find(".//startTime").text[11:16]
+    sar_instrument_mode = root.find(".//mode").text
+    # sar_polarizations = root.find(".//polarisation").text
+    sar_looks_range = "1"
+    sar_looks_azimuth = "1"
+    # sar_observation_direction = "right" root.find(".//").find("pass").text # left or right
+    sar_pixel_spacing_range = root.find(".//rangePixelSpacing").text
+    sar_pixel_spacing_azimuth = root.find(".//azimuthPixelSpacing").text
+
+    # use equation to calculate the resolution
+    # eq: resolution_range = c/(2Bandwidth)
+    # eq: resolution_azimuth = antenna_length/2
+        # c = 3e8
+    # sar_resolution_range = c/2/float(root.find(".//rangeProcessing").find("totalBandwidth").text.strip('"'))  # ?? total bandwidth??
+    # sar_resolution_azimuth = root.find(".//").text  # ?? antenna length ???
+
+    # sar_processing_level = root.find(".//").text
+    # sar_processing_lineage = root.find(".//").text
+    sar_processing_software = "doris 5"
+    # sar_proj_epsg = root.find(".//").text
+    # sar_orbit_state = root.find(".//").text
+    sar_absolute_orbit = root.find(".//absoluteOrbitNumber").text
+    sar_relative_orbit = (int(sar_absolute_orbit)-73) % 175 +1
+    # sci_publications = root.find(".//").text
+    sar_view_azimuth = root.find(".//pass").text
+    sar_view_incidence_angle = root.find(".//incidenceAngleMidSwath").text
+    # sar_insar_master_time = # read from input file
+    sar_SLC_crop = [500, 1440, 16000, 18350] # read from input file
+
+
+    Global_attr = {'processing_level':'L1', 'date_created': '2023-09-23', 'creator_name': 'Xu Zhang', 'creator_email': 'x.zhang-7@utwente.nl', \
[…]
