### 2025-04-25 #7532 @LC-SAR
> Hi [@LC-SAR](https://github.com/LC-SAR), just checking in on the status of the updates. I imagine it is a lot of work to make the changes described above and it may take some time. I just want to re-iterate that if we close this submission you are free to re-submit to JOSS at any time. There won't be any penalty to having already submitted. In that case you can take as long as you want to make updates to the repository. If you don't want to go that route and still like to make updates to this submission please do so by April 28th. We have many waitlisted submissions that would also like to receive reviewer feedback.
>
> If you have any questions please feel free to ask.

Dear @rwegener2, @philippemiron and @shahchiragh, we managed to create 'tests' folder (https://github.com/AlignSAR/alignSAR/tree/main/tests) that provides scripts to test all our proposed scripts. We make these test scripts (semi)automatic. The main script for testing is https://github.com/AlignSAR/alignSAR/blob/main/tests/AlignSAR_test_main.py, and the instruction to run the test is https://github.com/AlignSAR/alignSAR/blob/main/tests/README . We have tested all these scripts internally. Please check and test them. We hope the 'tests' work smoothly for you. Note that the volume of the sample data we use for testing is a bit big, as our scripts do the batch processing for a series of SAR images (7 for our case).  Looking forward to your feedback soon.
